Crate esp32s2

Peripheral access API for ESP32-S2 microcontrollers (generated using svd2rust v0.31.5 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::i2c0 as i2c1;

pub use self::spi0 as spi1;

pub use self::spi0 as spi2;

pub use self::spi0 as spi3;

pub use self::spi0 as spi4;

pub use self::timg0 as timg1;

pub use self::uart0 as uart1;

Modules

aes

AES (Advanced Encryption Standard) Accelerator

apb_saradc

SAR (Successive Approximation Register) Analog-to-Digital Converter

bb

BB Peripheral

dedicated_gpio

DEDICATED_GPIO Peripheral

ds

Digital Signature

efuse

eFuse Controller

extmem

External Memory

generic

Common register and bit access and modify traits

gpio

General Purpose Input/Output

gpio_sd

Sigma-Delta Modulation

hmac

HMAC (Hash-based Message Authentication Code) Accelerator

i2c0

I2C (Inter-Integrated Circuit) Controller 0

i2s0

I2S (Inter-IC Sound) Controller 0

interrupt_core0

Interrupt Controller (Core 0)

io_mux

Input/Output Multiplexer

ledc

LED Control PWM (Pulse Width Modulation)

pcnt

Pulse Count Controller

pms

Permissions Controller

rmt

Remote Control

rng

Hardware Random Number Generator

rsa

RSA (Rivest Shamir Adleman) Accelerator

rtc_cntl

Real-Time Clock Control

rtc_i2c

Low-power I2C (Inter-Integrated Circuit) Controller

rtc_io

Low-power Input/Output

sens

SENS Peripheral

sha

SHA (Secure Hash Algorithm) Accelerator

spi0

SPI (Serial Peripheral Interface) Controller 0

syscon

SYSCON Peripheral

system

System Configuration Registers

systimer

System Timer

timg0

Timer Group 0

twai0

Two-Wire Automotive Interface

uart0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

uhci0

Universal Host Controller Interface 0

usb0

USB OTG (On-The-Go)

usb_wrap

USB_WRAP Peripheral

xts_aes

XTS-AES-128 Flash Encryption

Structs

AES

AES (Advanced Encryption Standard) Accelerator

APB_SARADC

SAR (Successive Approximation Register) Analog-to-Digital Converter

BB

BB Peripheral

DEDICATED_GPIO

DEDICATED_GPIO Peripheral

DS

Digital Signature

EFUSE

eFuse Controller

EXTMEM

External Memory

GPIO

General Purpose Input/Output

GPIO_SD

Sigma-Delta Modulation

HMAC

HMAC (Hash-based Message Authentication Code) Accelerator

I2C0

I2C (Inter-Integrated Circuit) Controller 0

I2C1

I2C (Inter-Integrated Circuit) Controller 1

I2S0

I2S (Inter-IC Sound) Controller 0

INTERRUPT_CORE0

Interrupt Controller (Core 0)

IO_MUX

Input/Output Multiplexer

LEDC

LED Control PWM (Pulse Width Modulation)

PCNT

Pulse Count Controller

PMS

Permissions Controller

Peripherals

All the peripherals.

RMT

Remote Control

RNG

Hardware Random Number Generator

RSA

RSA (Rivest Shamir Adleman) Accelerator

RTC_CNTL

Real-Time Clock Control

RTC_I2C

Low-power I2C (Inter-Integrated Circuit) Controller

RTC_IO

Low-power Input/Output

SENS

SENS Peripheral

SHA

SHA (Secure Hash Algorithm) Accelerator

SPI0

SPI (Serial Peripheral Interface) Controller 0

SPI1

SPI (Serial Peripheral Interface) Controller 1

SPI2

SPI (Serial Peripheral Interface) Controller 2

SPI3

SPI (Serial Peripheral Interface) Controller 3

SPI4

SPI (Serial Peripheral Interface) Controller 4

SYSCON

SYSCON Peripheral

SYSTEM

System Configuration Registers

SYSTIMER

System Timer

TIMG0

Timer Group 0

TIMG1

Timer Group 1

TWAI0

Two-Wire Automotive Interface

TryFromInterruptError

TryFromInterruptError

UART0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

UART1

UART (Universal Asynchronous Receiver-Transmitter) Controller 1

UHCI0

Universal Host Controller Interface 0

USB0

USB OTG (On-The-Go)

USB_WRAP

USB_WRAP Peripheral

XTS_AES

XTS-AES-128 Flash Encryption

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority