Type Alias esp32s2::dedicated_gpio::intr_raw::R

pub type R = R<INTR_RAW_SPEC>;

Register INTR_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn gpio0(&self) -> GPIO0_R

Bit 0 - This interrupt raw bit turns to high level when dedicated GPIO0 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio1(&self) -> GPIO1_R

Bit 1 - This interrupt raw bit turns to high level when dedicated GPIO1 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio2(&self) -> GPIO2_R

Bit 2 - This interrupt raw bit turns to high level when dedicated GPIO2 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio3(&self) -> GPIO3_R

Bit 3 - This interrupt raw bit turns to high level when dedicated GPIO3 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio4(&self) -> GPIO4_R

Bit 4 - This interrupt raw bit turns to high level when dedicated GPIO4 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio5(&self) -> GPIO5_R

Bit 5 - This interrupt raw bit turns to high level when dedicated GPIO5 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio6(&self) -> GPIO6_R

Bit 6 - This interrupt raw bit turns to high level when dedicated GPIO6 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.

pub fn gpio7(&self) -> GPIO7_R

Bit 7 - This interrupt raw bit turns to high level when dedicated GPIO7 has level/edge change configured by DEDIC_GPIO_INTR_RCGN_REG.