Module esp32s2::extmem::cache_dbg_int_ena

register description

Structs

CACHE_DBG_INT_ENA_SPEC

register description

Type Aliases

CACHE_DBG_EN_R

Field CACHE_DBG_EN reader - The bit is used to activate the cache track function. 1: enable, 0: disable.

CACHE_DBG_EN_W

Field CACHE_DBG_EN writer - The bit is used to activate the cache track function. 1: enable, 0: disable.

DBUS_ACS_MSK_DC_INT_ENA_R

Field DBUS_ACS_MSK_DC_INT_ENA reader - The bit is used to enable interrupt by cpu access dcache while the corresponding dbus is disabled which include speculative access.

DBUS_ACS_MSK_DC_INT_ENA_W

Field DBUS_ACS_MSK_DC_INT_ENA writer - The bit is used to enable interrupt by cpu access dcache while the corresponding dbus is disabled which include speculative access.

DBUS_CNT_OVF_INT_ENA_R

Field DBUS_CNT_OVF_INT_ENA reader - The bit is used to enable interrupt by dbus counter overflow.

DBUS_CNT_OVF_INT_ENA_W

Field DBUS_CNT_OVF_INT_ENA writer - The bit is used to enable interrupt by dbus counter overflow.

DCACHE_REJECT_INT_ENA_R

Field DCACHE_REJECT_INT_ENA reader - The bit is used to enable interrupt by authentication fail.

DCACHE_REJECT_INT_ENA_W

Field DCACHE_REJECT_INT_ENA writer - The bit is used to enable interrupt by authentication fail.

DCACHE_SET_LOCK_ILG_INT_ENA_R

Field DCACHE_SET_LOCK_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing lock registers of dcache while dcache is busy to issue lock,sync or pre-load operations.

DCACHE_SET_LOCK_ILG_INT_ENA_W

Field DCACHE_SET_LOCK_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing lock registers of dcache while dcache is busy to issue lock,sync or pre-load operations.

DCACHE_SET_PRELOAD_ILG_INT_ENA_R

Field DCACHE_SET_PRELOAD_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing preload registers of dcache while dcache is busy to issue lock,sync and pre-load operations.

DCACHE_SET_PRELOAD_ILG_INT_ENA_W

Field DCACHE_SET_PRELOAD_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing preload registers of dcache while dcache is busy to issue lock,sync and pre-load operations.

DCACHE_SET_SYNC_ILG_INT_ENA_R

Field DCACHE_SET_SYNC_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing sync registers of dcache while dcache is busy to issue lock,sync and pre-load operations.

DCACHE_SET_SYNC_ILG_INT_ENA_W

Field DCACHE_SET_SYNC_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing sync registers of dcache while dcache is busy to issue lock,sync and pre-load operations.

DCACHE_WRITE_FLASH_INT_ENA_R

Field DCACHE_WRITE_FLASH_INT_ENA reader - The bit is used to enable interrupt by dcache trying to write flash.

DCACHE_WRITE_FLASH_INT_ENA_W

Field DCACHE_WRITE_FLASH_INT_ENA writer - The bit is used to enable interrupt by dcache trying to write flash.

DC_PRELOAD_SIZE_FAULT_INT_ENA_R

Field DC_PRELOAD_SIZE_FAULT_INT_ENA reader - The bit is used to enable interrupt by manual pre-load configurations fault.

DC_PRELOAD_SIZE_FAULT_INT_ENA_W

Field DC_PRELOAD_SIZE_FAULT_INT_ENA writer - The bit is used to enable interrupt by manual pre-load configurations fault.

DC_SYNC_SIZE_FAULT_INT_ENA_R

Field DC_SYNC_SIZE_FAULT_INT_ENA reader - The bit is used to enable interrupt by manual sync configurations fault.

DC_SYNC_SIZE_FAULT_INT_ENA_W

Field DC_SYNC_SIZE_FAULT_INT_ENA writer - The bit is used to enable interrupt by manual sync configurations fault.

IBUS_ACS_MSK_IC_INT_ENA_R

Field IBUS_ACS_MSK_IC_INT_ENA reader - The bit is used to enable interrupt by cpu access icache while the corresponding ibus is disabled which include speculative access.

IBUS_ACS_MSK_IC_INT_ENA_W

Field IBUS_ACS_MSK_IC_INT_ENA writer - The bit is used to enable interrupt by cpu access icache while the corresponding ibus is disabled which include speculative access.

IBUS_CNT_OVF_INT_ENA_R

Field IBUS_CNT_OVF_INT_ENA reader - The bit is used to enable interrupt by ibus counter overflow.

IBUS_CNT_OVF_INT_ENA_W

Field IBUS_CNT_OVF_INT_ENA writer - The bit is used to enable interrupt by ibus counter overflow.

ICACHE_REJECT_INT_ENA_R

Field ICACHE_REJECT_INT_ENA reader - The bit is used to enable interrupt by authentication fail.

ICACHE_REJECT_INT_ENA_W

Field ICACHE_REJECT_INT_ENA writer - The bit is used to enable interrupt by authentication fail.

ICACHE_SET_LOCK_ILG_INT_ENA_R

Field ICACHE_SET_LOCK_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing lock registers of icache while icache is busy to issue lock,sync or pre-load operations.

ICACHE_SET_LOCK_ILG_INT_ENA_W

Field ICACHE_SET_LOCK_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing lock registers of icache while icache is busy to issue lock,sync or pre-load operations.

ICACHE_SET_PRELOAD_ILG_INT_ENA_R

Field ICACHE_SET_PRELOAD_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing preload registers of icache while icache is busy to issue lock,sync and pre-load operations.

ICACHE_SET_PRELOAD_ILG_INT_ENA_W

Field ICACHE_SET_PRELOAD_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing preload registers of icache while icache is busy to issue lock,sync and pre-load operations.

ICACHE_SET_SYNC_ILG_INT_ENA_R

Field ICACHE_SET_SYNC_ILG_INT_ENA reader - The bit is used to enable interrupt by illegal writing sync registers of icache while icache is busy to issue lock,sync and pre-load operations.

ICACHE_SET_SYNC_ILG_INT_ENA_W

Field ICACHE_SET_SYNC_ILG_INT_ENA writer - The bit is used to enable interrupt by illegal writing sync registers of icache while icache is busy to issue lock,sync and pre-load operations.

IC_PRELOAD_SIZE_FAULT_INT_ENA_R

Field IC_PRELOAD_SIZE_FAULT_INT_ENA reader - The bit is used to enable interrupt by manual pre-load configurations fault.

IC_PRELOAD_SIZE_FAULT_INT_ENA_W

Field IC_PRELOAD_SIZE_FAULT_INT_ENA writer - The bit is used to enable interrupt by manual pre-load configurations fault.

IC_SYNC_SIZE_FAULT_INT_ENA_R

Field IC_SYNC_SIZE_FAULT_INT_ENA reader - The bit is used to enable interrupt by manual sync configurations fault.

IC_SYNC_SIZE_FAULT_INT_ENA_W

Field IC_SYNC_SIZE_FAULT_INT_ENA writer - The bit is used to enable interrupt by manual sync configurations fault.

MMU_ENTRY_FAULT_INT_ENA_R

Field MMU_ENTRY_FAULT_INT_ENA reader - The bit is used to enable interrupt by mmu entry fault.

MMU_ENTRY_FAULT_INT_ENA_W

Field MMU_ENTRY_FAULT_INT_ENA writer - The bit is used to enable interrupt by mmu entry fault.

R

Register CACHE_DBG_INT_ENA reader

W

Register CACHE_DBG_INT_ENA writer