Type Alias esp32s2::extmem::pro_icache_autoload_cfg::R

pub type R = R<PRO_ICACHE_AUTOLOAD_CFG_SPEC>;

Register PRO_ICACHE_AUTOLOAD_CFG reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn pro_icache_autoload_mode(&self) -> PRO_ICACHE_AUTOLOAD_MODE_R

Bit 0 - Reserved.

pub fn pro_icache_autoload_step(&self) -> PRO_ICACHE_AUTOLOAD_STEP_R

Bits 1:2 - Reserved.

pub fn pro_icache_autoload_order(&self) -> PRO_ICACHE_AUTOLOAD_ORDER_R

Bit 3 - The bits are used to configure the direction of conditional pre-load operation. 1: descending, 0: ascending.

pub fn pro_icache_autoload_rqst(&self) -> PRO_ICACHE_AUTOLOAD_RQST_R

Bits 4:5 - The bits are used to configure trigger conditions for conditional pre-load. 0/3: cache miss, 1: cache hit, 2: both cache miss and hit.

pub fn pro_icache_autoload_size(&self) -> PRO_ICACHE_AUTOLOAD_SIZE_R

Bits 6:7 - The bits are used to configure the numbers of the cache block for the issuing conditional pre-load operation.

pub fn pro_icache_autoload_sct0_ena(&self) -> PRO_ICACHE_AUTOLOAD_SCT0_ENA_R

Bit 8 - The bits are used to enable the second section for conditional pre-load operation.

pub fn pro_icache_autoload_sct1_ena(&self) -> PRO_ICACHE_AUTOLOAD_SCT1_ENA_R

Bit 9 - The bits are used to enable the first section for conditional pre-load operation.