Module esp32s2::hmac

HMAC (Hash-based Message Authentication Code) Accelerator

Modules

date

Version control register

one_block

One block message register.

query_busy

The busy state of HMAC module

query_error

The matching result between key and purpose user configured

rd_result_

Hash result register %s

set_invalidate_ds

Invalidate digital signature result register

set_invalidate_jtag

Invalidate JTAG result register

set_message_end

HMAC message end register

set_message_ing

HMAC message continue register

set_message_one

HMAC one message control register

set_message_pad

Software padding register

set_para_finish

HMAC configuration completion register

set_para_key

HMAC key configuration register

set_para_purpose

HMAC parameter configuration register

set_result_finish

HMAC read result completion register

set_start

HMAC start control register

wr_message_

Message register %s

Structs

RegisterBlock

Register block

Type Aliases

DATE

DATE (rw) register accessor: Version control register

ONE_BLOCK

ONE_BLOCK (w) register accessor: One block message register.

QUERY_BUSY

QUERY_BUSY (r) register accessor: The busy state of HMAC module

QUERY_ERROR

QUERY_ERROR (r) register accessor: The matching result between key and purpose user configured

RD_RESULT_

RD_RESULT_ (r) register accessor: Hash result register %s

SET_INVALIDATE_DS

SET_INVALIDATE_DS (w) register accessor: Invalidate digital signature result register

SET_INVALIDATE_JTAG

SET_INVALIDATE_JTAG (w) register accessor: Invalidate JTAG result register

SET_MESSAGE_END

SET_MESSAGE_END (w) register accessor: HMAC message end register

SET_MESSAGE_ING

SET_MESSAGE_ING (w) register accessor: HMAC message continue register

SET_MESSAGE_ONE

SET_MESSAGE_ONE (w) register accessor: HMAC one message control register

SET_MESSAGE_PAD

SET_MESSAGE_PAD (w) register accessor: Software padding register

SET_PARA_FINISH

SET_PARA_FINISH (w) register accessor: HMAC configuration completion register

SET_PARA_KEY

SET_PARA_KEY (w) register accessor: HMAC key configuration register

SET_PARA_PURPOSE

SET_PARA_PURPOSE (w) register accessor: HMAC parameter configuration register

SET_RESULT_FINISH

SET_RESULT_FINISH (w) register accessor: HMAC read result completion register

SET_START

SET_START (w) register accessor: HMAC start control register

WR_MESSAGE_

WR_MESSAGE_ (w) register accessor: Message register %s