Module esp32s2::i2c0::sr

Describe I2C work status

Structs

SR_SPEC

Describe I2C work status

Type Aliases

ARB_LOST_R

Field ARB_LOST reader - When the I2C controller loses control of SCL line, this register changes to 1.

BUS_BUSY_R

Field BUS_BUSY reader - 1: the I2C bus is busy transferring data. 0: the I2C bus is in idle state.

BYTE_TRANS_R

Field BYTE_TRANS reader - This field changes to 1 when one byte is transferred.

R

Register SR reader

RESP_REC_R

Field RESP_REC reader - The received ACK value in master mode or slave mode. 0: ACK. 1: NACK.

RXFIFO_CNT_R

Field RXFIFO_CNT reader - This field represents the amount of data needed to be sent.

SCL_MAIN_STATE_LAST_R

Field SCL_MAIN_STATE_LAST reader - This field indicates the states of the I2C module state machine. 0: Idle. 1: Address shift. 2: ACK address. 3: RX data. 4: TX data. 5: Send ACK. 6: Wait ACK

SCL_STATE_LAST_R

Field SCL_STATE_LAST reader - This field indicates the states of the state machine used to produce SCL. 0: Idle. 1: Start. 2: Negative edge. 3: Low. 4: Positive edge. 5: High. 6: Stop

SLAVE_ADDRESSED_R

Field SLAVE_ADDRESSED reader - When configured as an I2C Slave, and the address sent by the master is equal to the address of the slave, then this bit will be of high level.

SLAVE_RW_R

Field SLAVE_RW reader - When in slave mode, 1: master reads from slave. 0: master writes to slave.

STRETCH_CAUSE_R

Field STRETCH_CAUSE reader - The cause of stretching SCL low in slave mode. 0: stretching SCL low at the beginning of I2C read data state. 1: stretching SCL low when I2C TX FIFO is empty in slave mode. 2: stretching SCL low when I2C RX FIFO is full in slave mode.

TIME_OUT_R

Field TIME_OUT reader - When the I2C controller takes more than I2C_TIME_OUT clocks to receive a data bit, this field changes to 1.

TXFIFO_CNT_R

Field TXFIFO_CNT reader - This field stores the amount of received data in RAM.