Module esp32s2::i2s0::conf

I2S configuration register

Structs

CONF_SPEC

I2S configuration register

Type Aliases

PRE_REQ_EN_R

Field PRE_REQ_EN reader - Set this bit to enable I2S to prepare data earlier.

PRE_REQ_EN_W

Field PRE_REQ_EN writer - Set this bit to enable I2S to prepare data earlier.

R

Register CONF reader

RX_BIG_ENDIAN_R

Field RX_BIG_ENDIAN reader - I2S RX byte endianness.

RX_BIG_ENDIAN_W

Field RX_BIG_ENDIAN writer - I2S RX byte endianness.

RX_DMA_EQUAL_R

Field RX_DMA_EQUAL reader - 1: Data in left channel is equal to data in right channel. 0: Data in left channel is not equal to data in right channel.

RX_DMA_EQUAL_W

Field RX_DMA_EQUAL writer - 1: Data in left channel is equal to data in right channel. 0: Data in left channel is not equal to data in right channel.

RX_FIFO_RESET_ST_R

Field RX_FIFO_RESET_ST reader - I2S RX FIFO reset status. 1: I2S_RX_FIFO_RESET is not completed. 0: I2S_RX_FIFO_RESET is completed.

RX_FIFO_RESET_W

Field RX_FIFO_RESET writer - Set this bit to reset RX FIFO.

RX_LSB_FIRST_DMA_R

Field RX_LSB_FIRST_DMA reader - 1:the data in DMA/APB transform from low bits. 0:the data from DMA/APB transform from high bits.

RX_LSB_FIRST_DMA_W

Field RX_LSB_FIRST_DMA writer - 1:the data in DMA/APB transform from low bits. 0:the data from DMA/APB transform from high bits.

RX_MONO_R

Field RX_MONO reader - Set this bit to enable receiver in mono mode.

RX_MONO_W

Field RX_MONO writer - Set this bit to enable receiver in mono mode.

RX_MSB_RIGHT_R

Field RX_MSB_RIGHT reader - Set this bit to place right channel data at the MSB in RX FIFO.

RX_MSB_RIGHT_W

Field RX_MSB_RIGHT writer - Set this bit to place right channel data at the MSB in RX FIFO.

RX_MSB_SHIFT_R

Field RX_MSB_SHIFT reader - Set this bit to enable receiver in Phillips standard mode.

RX_MSB_SHIFT_W

Field RX_MSB_SHIFT writer - Set this bit to enable receiver in Phillips standard mode.

RX_RESET_ST_R

Field RX_RESET_ST reader - I2S RX reset status. 1: I2S_RX_RESET is not completed. 0: I2S_RX_RESET is completed.

RX_RESET_W

Field RX_RESET writer - Set this bit to reset receiver.

RX_RIGHT_FIRST_R

Field RX_RIGHT_FIRST reader - Set this bit to receive right channel data first.

RX_RIGHT_FIRST_W

Field RX_RIGHT_FIRST writer - Set this bit to receive right channel data first.

RX_SHORT_SYNC_R

Field RX_SHORT_SYNC reader - Set this bit to enable receiver in PCM standard mode.

RX_SHORT_SYNC_W

Field RX_SHORT_SYNC writer - Set this bit to enable receiver in PCM standard mode.

RX_SLAVE_MOD_R

Field RX_SLAVE_MOD reader - Set this bit to enable slave receiver mode.

RX_SLAVE_MOD_W

Field RX_SLAVE_MOD writer - Set this bit to enable slave receiver mode.

RX_START_R

Field RX_START reader - Set this bit to start receiving data.

RX_START_W

Field RX_START writer - Set this bit to start receiving data.

SIG_LOOPBACK_R

Field SIG_LOOPBACK reader - Enable signal loopback mode with transmitter module and receiver module sharing the same WS and BCK signals.

SIG_LOOPBACK_W

Field SIG_LOOPBACK writer - Enable signal loopback mode with transmitter module and receiver module sharing the same WS and BCK signals.

TX_BIG_ENDIAN_R

Field TX_BIG_ENDIAN reader - I2S TX byte endianness.

TX_BIG_ENDIAN_W

Field TX_BIG_ENDIAN writer - I2S TX byte endianness.

TX_DMA_EQUAL_R

Field TX_DMA_EQUAL reader - 1: Data in left channel is equal to data in right channel. 0: Data in left channel is not equal to data in right channel.

TX_DMA_EQUAL_W

Field TX_DMA_EQUAL writer - 1: Data in left channel is equal to data in right channel. 0: Data in left channel is not equal to data in right channel.

TX_FIFO_RESET_ST_R

Field TX_FIFO_RESET_ST reader - I2S TX FIFO reset status. 1: I2S_TX_FIFO_RESET is not completed. 0: I2S_TX_FIFO_RESET is completed.

TX_FIFO_RESET_W

Field TX_FIFO_RESET writer - Set this bit to reset TX FIFO.

TX_LSB_FIRST_DMA_R

Field TX_LSB_FIRST_DMA reader - 1:the data in DMA/APB transform from low bits. 0:the data from DMA/APB transform from high bits.

TX_LSB_FIRST_DMA_W

Field TX_LSB_FIRST_DMA writer - 1:the data in DMA/APB transform from low bits. 0:the data from DMA/APB transform from high bits.

TX_MONO_R

Field TX_MONO reader - Set this bit to enable transmitter in mono mode.

TX_MONO_W

Field TX_MONO writer - Set this bit to enable transmitter in mono mode.

TX_MSB_RIGHT_R

Field TX_MSB_RIGHT reader - Set this bit to place right channel data at the MSB in TX FIFO.

TX_MSB_RIGHT_W

Field TX_MSB_RIGHT writer - Set this bit to place right channel data at the MSB in TX FIFO.

TX_MSB_SHIFT_R

Field TX_MSB_SHIFT reader - Set this bit to enable transmitter in Phillips standard mode.

TX_MSB_SHIFT_W

Field TX_MSB_SHIFT writer - Set this bit to enable transmitter in Phillips standard mode.

TX_RESET_ST_R

Field TX_RESET_ST reader - I2S TX reset status. 1: I2S_TX_RESET is not completed. 0: I2S_TX_RESET is completed.

TX_RESET_W

Field TX_RESET writer - Set this bit to reset transmitter.

TX_RIGHT_FIRST_R

Field TX_RIGHT_FIRST reader - Set this bit to transmit right channel data first.

TX_RIGHT_FIRST_W

Field TX_RIGHT_FIRST writer - Set this bit to transmit right channel data first.

TX_SHORT_SYNC_R

Field TX_SHORT_SYNC reader - Set this bit to enable transmitter in PCM standard mode.

TX_SHORT_SYNC_W

Field TX_SHORT_SYNC writer - Set this bit to enable transmitter in PCM standard mode.

TX_SLAVE_MOD_R

Field TX_SLAVE_MOD reader - Set this bit to enable slave transmitter mode.

TX_SLAVE_MOD_W

Field TX_SLAVE_MOD writer - Set this bit to enable slave transmitter mode.

TX_START_R

Field TX_START reader - Set this bit to start transmitting data.

TX_START_W

Field TX_START writer - Set this bit to start transmitting data.

W

Register CONF writer