Module esp32s2::i2s0::lc_conf

I2S DMA configuration register

Structs

LC_CONF_SPEC

I2S DMA configuration register

Type Aliases

AHBM_FIFO_RST_R

Field AHBM_FIFO_RST reader - Set this bit to reset AHB interface cmdFIFO of DMA. Set this bit before the DMA configuration.

AHBM_FIFO_RST_W

Field AHBM_FIFO_RST writer - Set this bit to reset AHB interface cmdFIFO of DMA. Set this bit before the DMA configuration.

AHBM_RST_R

Field AHBM_RST reader - Set this bit to reset AHB interface of DMA. Set this bit before the DMA configuration.

AHBM_RST_W

Field AHBM_RST writer - Set this bit to reset AHB interface of DMA. Set this bit before the DMA configuration.

CHECK_OWNER_R

Field CHECK_OWNER reader - Set this bit to enable check owner bit by hardware.

CHECK_OWNER_W

Field CHECK_OWNER writer - Set this bit to enable check owner bit by hardware.

EXT_MEM_BK_SIZE_R

Field EXT_MEM_BK_SIZE reader - DMA access external memory block size. 0: 16 bytes. 1: 32 bytes. 2: 64 bytes. 3: reserved.

EXT_MEM_BK_SIZE_W

Field EXT_MEM_BK_SIZE writer - DMA access external memory block size. 0: 16 bytes. 1: 32 bytes. 2: 64 bytes. 3: reserved.

INDSCR_BURST_EN_R

Field INDSCR_BURST_EN reader - DMA inlink descriptor transfer mode configuration bit. 1: Prepare inlink descriptor with burst mode. 0: Prepare inlink descriptor with byte mode.

INDSCR_BURST_EN_W

Field INDSCR_BURST_EN writer - DMA inlink descriptor transfer mode configuration bit. 1: Prepare inlink descriptor with burst mode. 0: Prepare inlink descriptor with byte mode.

IN_LOOP_TEST_R

Field IN_LOOP_TEST reader - Set this bit to loop test outlink.

IN_LOOP_TEST_W

Field IN_LOOP_TEST writer - Set this bit to loop test outlink.

IN_RST_R

Field IN_RST reader - Set this bit to reset in-DMA FSM. Set this bit before the DMA configuration.

IN_RST_W

Field IN_RST writer - Set this bit to reset in-DMA FSM. Set this bit before the DMA configuration.

MEM_TRANS_EN_R

Field MEM_TRANS_EN reader - Reserved.

MEM_TRANS_EN_W

Field MEM_TRANS_EN writer - Reserved.

OUTDSCR_BURST_EN_R

Field OUTDSCR_BURST_EN reader - DMA outlink descriptor transfer mode configuration bit. 1: Prepare outlink descriptor with burst mode. 0: Prepare outlink descriptor with byte mode.

OUTDSCR_BURST_EN_W

Field OUTDSCR_BURST_EN writer - DMA outlink descriptor transfer mode configuration bit. 1: Prepare outlink descriptor with burst mode. 0: Prepare outlink descriptor with byte mode.

OUT_AUTO_WRBACK_R

Field OUT_AUTO_WRBACK reader - Set this bit to enable outlink-written-back automatically when out buffer is transmitted done.

OUT_AUTO_WRBACK_W

Field OUT_AUTO_WRBACK writer - Set this bit to enable outlink-written-back automatically when out buffer is transmitted done.

OUT_DATA_BURST_EN_R

Field OUT_DATA_BURST_EN reader - Transmitter data transfer mode configuration bit. 1: Prepare out data with burst mode. 0: Prepare out data with byte mode.

OUT_DATA_BURST_EN_W

Field OUT_DATA_BURST_EN writer - Transmitter data transfer mode configuration bit. 1: Prepare out data with burst mode. 0: Prepare out data with byte mode.

OUT_EOF_MODE_R

Field OUT_EOF_MODE reader - DMA out EOF flag generation mode. 1: When DMA has popped all data from the FIFO. 0: When AHB has pushed all data to the FIFO.

OUT_EOF_MODE_W

Field OUT_EOF_MODE writer - DMA out EOF flag generation mode. 1: When DMA has popped all data from the FIFO. 0: When AHB has pushed all data to the FIFO.

OUT_LOOP_TEST_R

Field OUT_LOOP_TEST reader - Set this bit to loop test inlink.

OUT_LOOP_TEST_W

Field OUT_LOOP_TEST writer - Set this bit to loop test inlink.

OUT_NO_RESTART_CLR_R

Field OUT_NO_RESTART_CLR reader - Reserved.

OUT_NO_RESTART_CLR_W

Field OUT_NO_RESTART_CLR writer - Reserved.

OUT_RST_R

Field OUT_RST reader - Set this bit to reset out-DMA FSM. Set this bit before the DMA configuration.

OUT_RST_W

Field OUT_RST writer - Set this bit to reset out-DMA FSM. Set this bit before the DMA configuration.

R

Register LC_CONF reader

W

Register LC_CONF writer