Module esp32s2::rtc_cntl

Real-Time Clock Control

Modules

ana_conf

Configures the power options for I2C and PLLA

bias_conf

configure power register

brown_out

Brownout configuration register

clk_conf

RTC clock configuration register

cocpu_ctrl

ULP-RISCV configuration register

cpu_period_conf

CPU sel option

date

diag0

debug register

dig_iso

Digital system ISO configuration register

dig_pad_hold

Configures the hold option for digital GPIOs

dig_pwc

Digital system power configuraiton register

ext_wakeup1

EXT1 wakeup configuration register

ext_wakeup1_status

EXT1 wakeup source register

ext_wakeup_conf

GPIO wakeup configuration register

ext_xtl_conf

32 kHz crystal oscillator configuration register

int_clr_rtc

RTC interrupt clear register

int_ena_rtc

RTC interrupt enabling register

int_raw_rtc

RTC interrupt raw register

int_st_rtc

RTC interrupt state register

low_power_st

RTC main state machine status register

options0

Sets the power options of crystal and PLL clocks, and initiates reset by software

options1

RTC option register

pad_hold

Configures the hold options for RTC GPIOs

pwc

RTC power configuraiton register

reg

RTC/DIG regulator configuration register

reset_state

Indicates the CPU reset source. For more information about the reset cause, please refer to Table \ref{table:resetreasons} in Chapter \ref{module:ResetandClock} \textit{ ameref{module:ResetandClock}}.

sdio_act_conf

configure sdio active register

sdio_conf

configure vddsdio register

slow_clk_conf

RTC slow clock configuration register

slp_reject_cause

Stores the reject-to-sleep cause.

slp_reject_conf

Configures sleep / reject options

slp_timer0

RTC timer threshold register 0

slp_timer1

RTC timer threshold register 1

slp_wakeup_cause

Stores the sleep-to-wakeup cause.

state0

Configures the sleep / reject / wakeup state

store0

Reservation register 0

store1

Reservation register 1

store2

Reservation register 2

store3

Reservation register 3

store4

Reservation register 4

store5

Reservation register 5

store6

Reservation register 6

store7

Reservation register 7

sw_cpu_stall

CPU stall configuration register

swd_conf

Super watchdog configuration register

swd_wprotect

Super watchdog write protection configuration register

time_high0

Stores the higher 16 bits of RTC timer 0

time_high1

Stores the higher 16 bits of RTC timer 1

time_low0

Stores the lower 32 bits of RTC timer 0.

time_low1

Stores the lower 32 bits of RTC timer 1

time_update

RTC timer update control register

timer1

Configures CPU stall options

timer2

Configures RTC slow clock and touch controller

timer3

configure some wait time for power on

timer4

configure some wait time for power on

timer5

Configures the minimal sleep cycles

timer6

Configure minimal sleep cycles register

touch_approach

Configure touch approach settings

touch_ctrl1

Touch control register

touch_ctrl2

Touch control register

touch_filter_ctrl

Configure touch filter settings

touch_scan_ctrl

Configure touch scan settings

touch_slp_thres

Configure the settings of touch sleep pad

touch_timeout_ctrl

Configure touch timeout settings

ulp_cp_ctrl

ULP-FSM configuration register

ulp_cp_timer

Configure coprocessor timer

ulp_cp_timer_1

Configure sleep cycle of the timer

usb_conf

configure usb control register

wakeup_state

Wakeup bitmap enabling register

wdtconfig0

RTC watchdog configuration register

wdtconfig1

Configures the hold time of RTC watchdog at level 1

wdtconfig2

Configures the hold time of RTC watchdog at level 2

wdtconfig3

Configures the hold time of RTC watchdog at level 3

wdtconfig4

Configures the hold time of RTC watchdog at level 4

wdtfeed

RTC watchdog SW feed configuration register

wdtwprotect

RTC watchdog write protection configuration register

xtal32k_clk_factor

Configures the divider factor for the backup clock of 32 kHz crystal oscillator

xtal32k_conf

32 kHz crystal oscillator configuration register

Structs

RegisterBlock

Register block

Type Aliases

ANA_CONF

ANA_CONF (rw) register accessor: Configures the power options for I2C and PLLA

BIAS_CONF

BIAS_CONF (rw) register accessor: configure power register

BROWN_OUT

BROWN_OUT (rw) register accessor: Brownout configuration register

CLK_CONF

CLK_CONF (rw) register accessor: RTC clock configuration register

COCPU_CTRL

COCPU_CTRL (rw) register accessor: ULP-RISCV configuration register

CPU_PERIOD_CONF

CPU_PERIOD_CONF (rw) register accessor: CPU sel option

DATE

DATE (rw) register accessor:

DIAG0

DIAG0 (r) register accessor: debug register

DIG_ISO

DIG_ISO (rw) register accessor: Digital system ISO configuration register

DIG_PAD_HOLD

DIG_PAD_HOLD (rw) register accessor: Configures the hold option for digital GPIOs

DIG_PWC

DIG_PWC (rw) register accessor: Digital system power configuraiton register

EXT_WAKEUP1

EXT_WAKEUP1 (rw) register accessor: EXT1 wakeup configuration register

EXT_WAKEUP1_STATUS

EXT_WAKEUP1_STATUS (r) register accessor: EXT1 wakeup source register

EXT_WAKEUP_CONF

EXT_WAKEUP_CONF (rw) register accessor: GPIO wakeup configuration register

EXT_XTL_CONF

EXT_XTL_CONF (rw) register accessor: 32 kHz crystal oscillator configuration register

INT_CLR_RTC

INT_CLR_RTC (w) register accessor: RTC interrupt clear register

INT_ENA_RTC

INT_ENA_RTC (rw) register accessor: RTC interrupt enabling register

INT_RAW_RTC

INT_RAW_RTC (r) register accessor: RTC interrupt raw register

INT_ST_RTC

INT_ST_RTC (r) register accessor: RTC interrupt state register

LOW_POWER_ST

LOW_POWER_ST (r) register accessor: RTC main state machine status register

OPTIONS0

OPTIONS0 (rw) register accessor: Sets the power options of crystal and PLL clocks, and initiates reset by software

OPTIONS1

OPTIONS1 (rw) register accessor: RTC option register

PAD_HOLD

PAD_HOLD (rw) register accessor: Configures the hold options for RTC GPIOs

PWC

PWC (rw) register accessor: RTC power configuraiton register

REG

REG (rw) register accessor: RTC/DIG regulator configuration register

RESET_STATE

RESET_STATE (rw) register accessor: Indicates the CPU reset source. For more information about the reset cause, please refer to Table \ref{table:resetreasons} in Chapter \ref{module:ResetandClock} \textit{ ameref{module:ResetandClock}}.

SDIO_ACT_CONF

SDIO_ACT_CONF (rw) register accessor: configure sdio active register

SDIO_CONF

SDIO_CONF (rw) register accessor: configure vddsdio register

SLOW_CLK_CONF

SLOW_CLK_CONF (rw) register accessor: RTC slow clock configuration register

SLP_REJECT_CAUSE

SLP_REJECT_CAUSE (r) register accessor: Stores the reject-to-sleep cause.

SLP_REJECT_CONF

SLP_REJECT_CONF (rw) register accessor: Configures sleep / reject options

SLP_TIMER0

SLP_TIMER0 (rw) register accessor: RTC timer threshold register 0

SLP_TIMER1

SLP_TIMER1 (rw) register accessor: RTC timer threshold register 1

SLP_WAKEUP_CAUSE

SLP_WAKEUP_CAUSE (r) register accessor: Stores the sleep-to-wakeup cause.

STATE0

STATE0 (rw) register accessor: Configures the sleep / reject / wakeup state

STORE0

STORE0 (rw) register accessor: Reservation register 0

STORE1

STORE1 (rw) register accessor: Reservation register 1

STORE2

STORE2 (rw) register accessor: Reservation register 2

STORE3

STORE3 (rw) register accessor: Reservation register 3

STORE4

STORE4 (rw) register accessor: Reservation register 4

STORE5

STORE5 (rw) register accessor: Reservation register 5

STORE6

STORE6 (rw) register accessor: Reservation register 6

STORE7

STORE7 (rw) register accessor: Reservation register 7

SWD_CONF

SWD_CONF (rw) register accessor: Super watchdog configuration register

SWD_WPROTECT

SWD_WPROTECT (rw) register accessor: Super watchdog write protection configuration register

SW_CPU_STALL

SW_CPU_STALL (rw) register accessor: CPU stall configuration register

TIMER1

TIMER1 (rw) register accessor: Configures CPU stall options

TIMER2

TIMER2 (rw) register accessor: Configures RTC slow clock and touch controller

TIMER3

TIMER3 (rw) register accessor: configure some wait time for power on

TIMER4

TIMER4 (rw) register accessor: configure some wait time for power on

TIMER5

TIMER5 (rw) register accessor: Configures the minimal sleep cycles

TIMER6

TIMER6 (rw) register accessor: Configure minimal sleep cycles register

TIME_HIGH0

TIME_HIGH0 (r) register accessor: Stores the higher 16 bits of RTC timer 0

TIME_HIGH1

TIME_HIGH1 (r) register accessor: Stores the higher 16 bits of RTC timer 1

TIME_LOW0

TIME_LOW0 (r) register accessor: Stores the lower 32 bits of RTC timer 0.

TIME_LOW1

TIME_LOW1 (r) register accessor: Stores the lower 32 bits of RTC timer 1

TIME_UPDATE

TIME_UPDATE (rw) register accessor: RTC timer update control register

TOUCH_APPROACH

TOUCH_APPROACH (rw) register accessor: Configure touch approach settings

TOUCH_CTRL1

TOUCH_CTRL1 (rw) register accessor: Touch control register

TOUCH_CTRL2

TOUCH_CTRL2 (rw) register accessor: Touch control register

TOUCH_FILTER_CTRL

TOUCH_FILTER_CTRL (rw) register accessor: Configure touch filter settings

TOUCH_SCAN_CTRL

TOUCH_SCAN_CTRL (rw) register accessor: Configure touch scan settings

TOUCH_SLP_THRES

TOUCH_SLP_THRES (rw) register accessor: Configure the settings of touch sleep pad

TOUCH_TIMEOUT_CTRL

TOUCH_TIMEOUT_CTRL (rw) register accessor: Configure touch timeout settings

ULP_CP_CTRL

ULP_CP_CTRL (rw) register accessor: ULP-FSM configuration register

ULP_CP_TIMER

ULP_CP_TIMER (rw) register accessor: Configure coprocessor timer

ULP_CP_TIMER_1

ULP_CP_TIMER_1 (rw) register accessor: Configure sleep cycle of the timer

USB_CONF

USB_CONF (rw) register accessor: configure usb control register

WAKEUP_STATE

WAKEUP_STATE (rw) register accessor: Wakeup bitmap enabling register

WDTCONFIG0

WDTCONFIG0 (rw) register accessor: RTC watchdog configuration register

WDTCONFIG1

WDTCONFIG1 (rw) register accessor: Configures the hold time of RTC watchdog at level 1

WDTCONFIG2

WDTCONFIG2 (rw) register accessor: Configures the hold time of RTC watchdog at level 2

WDTCONFIG3

WDTCONFIG3 (rw) register accessor: Configures the hold time of RTC watchdog at level 3

WDTCONFIG4

WDTCONFIG4 (rw) register accessor: Configures the hold time of RTC watchdog at level 4

WDTFEED

WDTFEED (w) register accessor: RTC watchdog SW feed configuration register

WDTWPROTECT

WDTWPROTECT (rw) register accessor: RTC watchdog write protection configuration register

XTAL32K_CLK_FACTOR

XTAL32K_CLK_FACTOR (rw) register accessor: Configures the divider factor for the backup clock of 32 kHz crystal oscillator

XTAL32K_CONF

XTAL32K_CONF (rw) register accessor: 32 kHz crystal oscillator configuration register