Type Alias esp32s2::rtc_cntl::clk_conf::R

pub type R = R<CLK_CONF_SPEC>;

Register CLK_CONF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ck8m_div_sel_vld(&self) -> CK8M_DIV_SEL_VLD_R

Bit 3 - Synchronizes the reg_ck8m_div_sel. Not that you have to invalidate the bus before switching clock, and validate the new clock.

pub fn ck8m_div(&self) -> CK8M_DIV_R

Bits 4:5 - Set the CK8M_D256_OUT divider. 00: divided by 128 01: divided by 256 10: divided by 512 11: divided by 1024.

pub fn enb_ck8m(&self) -> ENB_CK8M_R

Bit 6 - Set this bit to disable CK8M and CK8M_D256_OUT.

pub fn enb_ck8m_div(&self) -> ENB_CK8M_DIV_R

Bit 7 - Selects the CK8M_D256_OUT. 1: CK8M 0: CK8M divided by 256.

pub fn dig_xtal32k_en(&self) -> DIG_XTAL32K_EN_R

Bit 8 - Set this bit to enable CK_XTAL_32K clock for the digital core.

pub fn dig_clk8m_d256_en(&self) -> DIG_CLK8M_D256_EN_R

Bit 9 - Set this bit to enable CK8M_D256_OUT clock for the digital core.

pub fn dig_clk8m_en(&self) -> DIG_CLK8M_EN_R

Bit 10 - Set this bit to enable 8 MHz clock for the digital core.

pub fn ck8m_div_sel(&self) -> CK8M_DIV_SEL_R

Bits 12:14 - Stores the 8 MHz divider, which is reg_ck8m_div_sel + 1

pub fn xtal_force_nogating(&self) -> XTAL_FORCE_NOGATING_R

Bit 15 - Set this bit to force no gating to crystal during sleep

pub fn ck8m_force_nogating(&self) -> CK8M_FORCE_NOGATING_R

Bit 16 - Set this bit to disable force gating to 8 MHz crystal during sleep.

pub fn ck8m_dfreq(&self) -> CK8M_DFREQ_R

Bits 17:24 - CK8M_DFREQ

pub fn ck8m_force_pd(&self) -> CK8M_FORCE_PD_R

Bit 25 - Set this bit to FPD the 8 MHz clock.

pub fn ck8m_force_pu(&self) -> CK8M_FORCE_PU_R

Bit 26 - Set this bit to FPU the 8 MHz clock.

pub fn fast_clk_rtc_sel(&self) -> FAST_CLK_RTC_SEL_R

Bit 29 - Set this bit to select the RTC fast clock. 0: XTAL div 4, 1: CK8M.

pub fn ana_clk_rtc_sel(&self) -> ANA_CLK_RTC_SEL_R

Bits 30:31 - Set this bit to select the RTC slow clock. 0: 90K rtc_clk 1: 32k XTAL 2: 8md256.