Type Alias esp32s2::rtc_cntl::dig_iso::W

pub type W = W<DIG_ISO_SPEC>;

Register DIG_ISO writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn force_off(&mut self) -> FORCE_OFF_W<'_, DIG_ISO_SPEC>

Bit 7

pub fn force_on(&mut self) -> FORCE_ON_W<'_, DIG_ISO_SPEC>

Bit 8

pub fn clr_dg_pad_autohold(&mut self) -> CLR_DG_PAD_AUTOHOLD_W<'_, DIG_ISO_SPEC>

Bit 10 - Se this bit to clear the auto-hold enabler for the digital GPIOs.

pub fn dg_pad_autohold_en(&mut self) -> DG_PAD_AUTOHOLD_EN_W<'_, DIG_ISO_SPEC>

Bit 11 - Se this bit to allow the digital GPIOs to enter the autohold status.

pub fn dg_pad_force_noiso(&mut self) -> DG_PAD_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 12 - Set this bit to disable the force isolation to the digital GPIOs.

pub fn dg_pad_force_iso(&mut self) -> DG_PAD_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 13 - Set this bit to force isolate the digital GPIOs.

pub fn dg_pad_force_unhold(&mut self) -> DG_PAD_FORCE_UNHOLD_W<'_, DIG_ISO_SPEC>

Bit 14 - Set this bit the force unhold the digital GPIOs.

pub fn dg_pad_force_hold(&mut self) -> DG_PAD_FORCE_HOLD_W<'_, DIG_ISO_SPEC>

Bit 15 - Set this bit the force hold the digital GPIOs.

pub fn rom0_force_iso(&mut self) -> ROM0_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 16 - ROM force ISO

pub fn rom0_force_noiso(&mut self) -> ROM0_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 17 - ROM force no ISO

pub fn inter_ram0_force_iso( &mut self ) -> INTER_RAM0_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 18 - internal SRAM 0 force ISO

pub fn inter_ram0_force_noiso( &mut self ) -> INTER_RAM0_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 19 - internal SRAM 0 force no ISO

pub fn inter_ram1_force_iso( &mut self ) -> INTER_RAM1_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 20 - internal SRAM 1 force ISO

pub fn inter_ram1_force_noiso( &mut self ) -> INTER_RAM1_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 21 - internal SRAM 1 force no ISO

pub fn inter_ram2_force_iso( &mut self ) -> INTER_RAM2_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 22 - internal SRAM 2 force ISO

pub fn inter_ram2_force_noiso( &mut self ) -> INTER_RAM2_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 23 - internal SRAM 2 force no ISO

pub fn inter_ram3_force_iso( &mut self ) -> INTER_RAM3_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 24 - internal SRAM 3 force ISO

pub fn inter_ram3_force_noiso( &mut self ) -> INTER_RAM3_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 25 - internal SRAM 3 force no ISO

pub fn inter_ram4_force_iso( &mut self ) -> INTER_RAM4_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 26 - internal SRAM 4 force ISO

pub fn inter_ram4_force_noiso( &mut self ) -> INTER_RAM4_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 27 - internal SRAM 4 force no ISO

pub fn wifi_force_iso(&mut self) -> WIFI_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 28 - Set this bit to force isolate the Wi-Fi circuits.

pub fn wifi_force_noiso(&mut self) -> WIFI_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 29 - Set this bit to disable the force isolation to the Wi-Fi circuits.

pub fn dg_wrap_force_iso(&mut self) -> DG_WRAP_FORCE_ISO_W<'_, DIG_ISO_SPEC>

Bit 30 - Set this bit to force isolate the digital system.

pub fn dg_wrap_force_noiso(&mut self) -> DG_WRAP_FORCE_NOISO_W<'_, DIG_ISO_SPEC>

Bit 31 - Set this bit to disable the force isolation to the digital system.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual