Module esp32s2::rtc_cntl::ext_xtl_conf

32 kHz crystal oscillator configuration register

Structs

EXT_XTL_CONF_SPEC

32 kHz crystal oscillator configuration register

Type Aliases

DAC_XTAL_32K_R

Field DAC_XTAL_32K reader - DAC_XTAL_32K

DAC_XTAL_32K_W

Field DAC_XTAL_32K writer - DAC_XTAL_32K

DBUF_XTAL_32K_R

Field DBUF_XTAL_32K reader - 0: single-end buffer 1: differential buffer

DBUF_XTAL_32K_W

Field DBUF_XTAL_32K writer - 0: single-end buffer 1: differential buffer

DGM_XTAL_32K_R

Field DGM_XTAL_32K reader - xtal_32k gm control

DGM_XTAL_32K_W

Field DGM_XTAL_32K writer - xtal_32k gm control

DRES_XTAL_32K_R

Field DRES_XTAL_32K reader - DRES_XTAL_32K

DRES_XTAL_32K_W

Field DRES_XTAL_32K writer - DRES_XTAL_32K

ENCKINIT_XTAL_32K_R

Field ENCKINIT_XTAL_32K reader - Applies an internal clock to help the 32 kHz crystal to start.

ENCKINIT_XTAL_32K_W

Field ENCKINIT_XTAL_32K writer - Applies an internal clock to help the 32 kHz crystal to start.

R

Register EXT_XTL_CONF reader

W

Register EXT_XTL_CONF writer

WDT_STATE_R

Field WDT_STATE reader - Stores the status of the 32 kHz watchdog.

XPD_XTAL_32K_R

Field XPD_XTAL_32K reader - XPD_XTAL_32K

XPD_XTAL_32K_W

Field XPD_XTAL_32K writer - XPD_XTAL_32K

XTAL32K_AUTO_BACKUP_R

Field XTAL32K_AUTO_BACKUP reader - Set this bit to switch to the backup clock when the 32 kHz crystal is dead.

XTAL32K_AUTO_BACKUP_W

Field XTAL32K_AUTO_BACKUP writer - Set this bit to switch to the backup clock when the 32 kHz crystal is dead.

XTAL32K_AUTO_RESTART_R

Field XTAL32K_AUTO_RESTART reader - Set this bit to restart the 32 kHz crystal automatically when the 32 kHz crystal is dead.

XTAL32K_AUTO_RESTART_W

Field XTAL32K_AUTO_RESTART writer - Set this bit to restart the 32 kHz crystal automatically when the 32 kHz crystal is dead.

XTAL32K_AUTO_RETURN_R

Field XTAL32K_AUTO_RETURN reader - Set this bit to switch back to 32 kHz crystal when the 32 kHz crystal is restarted.

XTAL32K_AUTO_RETURN_W

Field XTAL32K_AUTO_RETURN writer - Set this bit to switch back to 32 kHz crystal when the 32 kHz crystal is restarted.

XTAL32K_EXT_CLK_FO_R

Field XTAL32K_EXT_CLK_FO reader - Set this bit to FPU the external clock of 32 kHz crystal.

XTAL32K_EXT_CLK_FO_W

Field XTAL32K_EXT_CLK_FO writer - Set this bit to FPU the external clock of 32 kHz crystal.

XTAL32K_GPIO_SEL_R

Field XTAL32K_GPIO_SEL reader - Selects the 32 kHz crystal clock. 0: selects the external 32 kHz clock. 1: selects clock from the RTC GPIO X32P_C.

XTAL32K_GPIO_SEL_W

Field XTAL32K_GPIO_SEL writer - Selects the 32 kHz crystal clock. 0: selects the external 32 kHz clock. 1: selects clock from the RTC GPIO X32P_C.

XTAL32K_WDT_CLK_FO_R

Field XTAL32K_WDT_CLK_FO reader - Set this bit to FPU the 32 kHz crystal watchdog clock.

XTAL32K_WDT_CLK_FO_W

Field XTAL32K_WDT_CLK_FO writer - Set this bit to FPU the 32 kHz crystal watchdog clock.

XTAL32K_WDT_EN_R

Field XTAL32K_WDT_EN reader - Set this bit to enable the 32 kHz crystal watchdog.

XTAL32K_WDT_EN_W

Field XTAL32K_WDT_EN writer - Set this bit to enable the 32 kHz crystal watchdog.

XTAL32K_WDT_RESET_R

Field XTAL32K_WDT_RESET reader - Set this bit to reset the 32 kHz crystal watchdog by SW.

XTAL32K_WDT_RESET_W

Field XTAL32K_WDT_RESET writer - Set this bit to reset the 32 kHz crystal watchdog by SW.

XTAL32K_XPD_FORCE_R

Field XTAL32K_XPD_FORCE reader - Set 1 to allow the software to FPD the 32 kHz crystal. Set 0 to allow the FSM to FPD the 32 kHz crystal. (R/W)

XTAL32K_XPD_FORCE_W

Field XTAL32K_XPD_FORCE writer - Set 1 to allow the software to FPD the 32 kHz crystal. Set 0 to allow the FSM to FPD the 32 kHz crystal. (R/W)

XTL_EXT_CTR_EN_R

Field XTL_EXT_CTR_EN reader - Enables the GPIO to power down the crystal oscillator.

XTL_EXT_CTR_EN_W

Field XTL_EXT_CTR_EN writer - Enables the GPIO to power down the crystal oscillator.

XTL_EXT_CTR_LV_R

Field XTL_EXT_CTR_LV reader - 0: powers down XTAL at high level 1: powers down XTAL at low level

XTL_EXT_CTR_LV_W

Field XTL_EXT_CTR_LV writer - 0: powers down XTAL at high level 1: powers down XTAL at low level