Module esp32s2::rtc_cntl::int_ena_rtc

RTC interrupt enabling register

Structs

INT_ENA_RTC_SPEC

RTC interrupt enabling register

Type Aliases

BROWN_OUT_INT_ENA_R

Field BROWN_OUT_INT_ENA reader - Enables the brown out interrupt.

BROWN_OUT_INT_ENA_W

Field BROWN_OUT_INT_ENA writer - Enables the brown out interrupt.

COCPU_INT_ENA_R

Field COCPU_INT_ENA reader - Enables the ULP-RISCV interrupt.

COCPU_INT_ENA_W

Field COCPU_INT_ENA writer - Enables the ULP-RISCV interrupt.

COCPU_TRAP_INT_ENA_R

Field COCPU_TRAP_INT_ENA reader - Enables interruption when the ULP-RISCV is trapped.

COCPU_TRAP_INT_ENA_W

Field COCPU_TRAP_INT_ENA writer - Enables interruption when the ULP-RISCV is trapped.

GLITCH_DET_INT_ENA_R

Field GLITCH_DET_INT_ENA reader - Enables interruption when a glitch is detected.

GLITCH_DET_INT_ENA_W

Field GLITCH_DET_INT_ENA writer - Enables interruption when a glitch is detected.

MAIN_TIMER_INT_ENA_R

Field MAIN_TIMER_INT_ENA reader - Enables the RTC main timer interrupt.

MAIN_TIMER_INT_ENA_W

Field MAIN_TIMER_INT_ENA writer - Enables the RTC main timer interrupt.

R

Register INT_ENA_RTC reader

SARADC1_INT_ENA_R

Field SARADC1_INT_ENA reader - Enables the SAR ADC 1 interrupt.

SARADC1_INT_ENA_W

Field SARADC1_INT_ENA writer - Enables the SAR ADC 1 interrupt.

SARADC2_INT_ENA_R

Field SARADC2_INT_ENA reader - Enables the SAR ADC 2 interrupt.

SARADC2_INT_ENA_W

Field SARADC2_INT_ENA writer - Enables the SAR ADC 2 interrupt.

SDIO_IDLE_INT_ENA_R

Field SDIO_IDLE_INT_ENA reader - Enables interruption when the SDIO idles.

SDIO_IDLE_INT_ENA_W

Field SDIO_IDLE_INT_ENA writer - Enables interruption when the SDIO idles.

SLP_REJECT_INT_ENA_R

Field SLP_REJECT_INT_ENA reader - Enables interruption when the chip rejects to go to sleep.

SLP_REJECT_INT_ENA_W

Field SLP_REJECT_INT_ENA writer - Enables interruption when the chip rejects to go to sleep.

SLP_WAKEUP_INT_ENA_R

Field SLP_WAKEUP_INT_ENA reader - Enables interruption when the chip wakes up from sleep.

SLP_WAKEUP_INT_ENA_W

Field SLP_WAKEUP_INT_ENA writer - Enables interruption when the chip wakes up from sleep.

SWD_INT_ENA_R

Field SWD_INT_ENA reader - Enables the super watchdog interrupt.

SWD_INT_ENA_W

Field SWD_INT_ENA writer - Enables the super watchdog interrupt.

TOUCH_ACTIVE_INT_ENA_R

Field TOUCH_ACTIVE_INT_ENA reader - Enables interruption when a touch is detected.

TOUCH_ACTIVE_INT_ENA_W

Field TOUCH_ACTIVE_INT_ENA writer - Enables interruption when a touch is detected.

TOUCH_DONE_INT_ENA_R

Field TOUCH_DONE_INT_ENA reader - Enables interruption upon the completion of a single touch.

TOUCH_DONE_INT_ENA_W

Field TOUCH_DONE_INT_ENA writer - Enables interruption upon the completion of a single touch.

TOUCH_INACTIVE_INT_ENA_R

Field TOUCH_INACTIVE_INT_ENA reader - Enables interruption when a touch is released.

TOUCH_INACTIVE_INT_ENA_W

Field TOUCH_INACTIVE_INT_ENA writer - Enables interruption when a touch is released.

TOUCH_SCAN_DONE_INT_ENA_R

Field TOUCH_SCAN_DONE_INT_ENA reader - Enables interruption upon the completion of a touch scanning.

TOUCH_SCAN_DONE_INT_ENA_W

Field TOUCH_SCAN_DONE_INT_ENA writer - Enables interruption upon the completion of a touch scanning.

TOUCH_TIMEOUT_INT_ENA_R

Field TOUCH_TIMEOUT_INT_ENA reader - Enables interruption when touch sensor times out.

TOUCH_TIMEOUT_INT_ENA_W

Field TOUCH_TIMEOUT_INT_ENA writer - Enables interruption when touch sensor times out.

TSENS_INT_ENA_R

Field TSENS_INT_ENA reader - Enables the touch sensor interrupt.

TSENS_INT_ENA_W

Field TSENS_INT_ENA writer - Enables the touch sensor interrupt.

ULP_CP_INT_ENA_R

Field ULP_CP_INT_ENA reader - Enables the ULP co-processor interrupt.

ULP_CP_INT_ENA_W

Field ULP_CP_INT_ENA writer - Enables the ULP co-processor interrupt.

W

Register INT_ENA_RTC writer

WDT_INT_ENA_R

Field WDT_INT_ENA reader - Enables the RTC watchdog interrupt.

WDT_INT_ENA_W

Field WDT_INT_ENA writer - Enables the RTC watchdog interrupt.

XTAL32K_DEAD_INT_ENA_R

Field XTAL32K_DEAD_INT_ENA reader - Enables interruption when the 32 kHz crystal is dead.

XTAL32K_DEAD_INT_ENA_W

Field XTAL32K_DEAD_INT_ENA writer - Enables interruption when the 32 kHz crystal is dead.