Type Alias esp32s2::rtc_cntl::int_ena_rtc::R

pub type R = R<INT_ENA_RTC_SPEC>;

Register INT_ENA_RTC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn slp_wakeup_int_ena(&self) -> SLP_WAKEUP_INT_ENA_R

Bit 0 - Enables interruption when the chip wakes up from sleep.

pub fn slp_reject_int_ena(&self) -> SLP_REJECT_INT_ENA_R

Bit 1 - Enables interruption when the chip rejects to go to sleep.

pub fn sdio_idle_int_ena(&self) -> SDIO_IDLE_INT_ENA_R

Bit 2 - Enables interruption when the SDIO idles.

pub fn wdt_int_ena(&self) -> WDT_INT_ENA_R

Bit 3 - Enables the RTC watchdog interrupt.

pub fn touch_scan_done_int_ena(&self) -> TOUCH_SCAN_DONE_INT_ENA_R

Bit 4 - Enables interruption upon the completion of a touch scanning.

pub fn ulp_cp_int_ena(&self) -> ULP_CP_INT_ENA_R

Bit 5 - Enables the ULP co-processor interrupt.

pub fn touch_done_int_ena(&self) -> TOUCH_DONE_INT_ENA_R

Bit 6 - Enables interruption upon the completion of a single touch.

pub fn touch_active_int_ena(&self) -> TOUCH_ACTIVE_INT_ENA_R

Bit 7 - Enables interruption when a touch is detected.

pub fn touch_inactive_int_ena(&self) -> TOUCH_INACTIVE_INT_ENA_R

Bit 8 - Enables interruption when a touch is released.

pub fn brown_out_int_ena(&self) -> BROWN_OUT_INT_ENA_R

Bit 9 - Enables the brown out interrupt.

pub fn main_timer_int_ena(&self) -> MAIN_TIMER_INT_ENA_R

Bit 10 - Enables the RTC main timer interrupt.

pub fn saradc1_int_ena(&self) -> SARADC1_INT_ENA_R

Bit 11 - Enables the SAR ADC 1 interrupt.

pub fn tsens_int_ena(&self) -> TSENS_INT_ENA_R

Bit 12 - Enables the touch sensor interrupt.

pub fn cocpu_int_ena(&self) -> COCPU_INT_ENA_R

Bit 13 - Enables the ULP-RISCV interrupt.

pub fn saradc2_int_ena(&self) -> SARADC2_INT_ENA_R

Bit 14 - Enables the SAR ADC 2 interrupt.

pub fn swd_int_ena(&self) -> SWD_INT_ENA_R

Bit 15 - Enables the super watchdog interrupt.

pub fn xtal32k_dead_int_ena(&self) -> XTAL32K_DEAD_INT_ENA_R

Bit 16 - Enables interruption when the 32 kHz crystal is dead.

pub fn cocpu_trap_int_ena(&self) -> COCPU_TRAP_INT_ENA_R

Bit 17 - Enables interruption when the ULP-RISCV is trapped.

pub fn touch_timeout_int_ena(&self) -> TOUCH_TIMEOUT_INT_ENA_R

Bit 18 - Enables interruption when touch sensor times out.

pub fn glitch_det_int_ena(&self) -> GLITCH_DET_INT_ENA_R

Bit 19 - Enables interruption when a glitch is detected.