Type Alias esp32s2::rtc_cntl::int_raw_rtc::R

pub type R = R<INT_RAW_RTC_SPEC>;

Register INT_RAW_RTC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn slp_wakeup_int_raw(&self) -> SLP_WAKEUP_INT_RAW_R

Bit 0 - Stores the raw interrupt triggered when the chip wakes up from sleep.

pub fn slp_reject_int_raw(&self) -> SLP_REJECT_INT_RAW_R

Bit 1 - Stores the raw interrupt triggered when the chip rejects to go to sleep.

pub fn sdio_idle_int_raw(&self) -> SDIO_IDLE_INT_RAW_R

Bit 2 - Stores the raw interrupt triggered when the SDIO idles.

pub fn wdt_int_raw(&self) -> WDT_INT_RAW_R

Bit 3 - Stores the raw RTC watchdog interrupt.

pub fn touch_scan_done_int_raw(&self) -> TOUCH_SCAN_DONE_INT_RAW_R

Bit 4 - Stores the raw interrupt triggered upon the completion of a touch scanning.

pub fn ulp_cp_int_raw(&self) -> ULP_CP_INT_RAW_R

Bit 5 - Stores the raw ULP co-processor interrupt.

pub fn touch_done_int_raw(&self) -> TOUCH_DONE_INT_RAW_R

Bit 6 - Stores the raw interrupt triggered upon the completion of a single touch.

pub fn touch_active_int_raw(&self) -> TOUCH_ACTIVE_INT_RAW_R

Bit 7 - Stores the raw interrupt triggered when a touch is detected.

pub fn touch_inactive_int_raw(&self) -> TOUCH_INACTIVE_INT_RAW_R

Bit 8 - Stores the raw interrupt triggered when a touch is released.

pub fn brown_out_int_raw(&self) -> BROWN_OUT_INT_RAW_R

Bit 9 - Stores the raw brown out interrupt.

pub fn main_timer_int_raw(&self) -> MAIN_TIMER_INT_RAW_R

Bit 10 - Stores the raw RTC main timer interrupt.

pub fn saradc1_int_raw(&self) -> SARADC1_INT_RAW_R

Bit 11 - Stores the raw SAR ADC 1 interrupt.

pub fn tsens_int_raw(&self) -> TSENS_INT_RAW_R

Bit 12 - Stores the raw touch sensor interrupt.

pub fn cocpu_int_raw(&self) -> COCPU_INT_RAW_R

Bit 13 - Stores the raw ULP-RISCV interrupt.

pub fn saradc2_int_raw(&self) -> SARADC2_INT_RAW_R

Bit 14 - Stores the raw SAR ADC 2 interrupt.

pub fn swd_int_raw(&self) -> SWD_INT_RAW_R

Bit 15 - Stores the raw super watchdog interrupt.

pub fn xtal32k_dead_int_raw(&self) -> XTAL32K_DEAD_INT_RAW_R

Bit 16 - Stores the raw interrupt triggered when the 32 kHz crystal is dead.

pub fn cocpu_trap_int_raw(&self) -> COCPU_TRAP_INT_RAW_R

Bit 17 - Stores the raw interrupt triggered when the ULP-RISCV is trapped.

pub fn touch_timeout_int_raw(&self) -> TOUCH_TIMEOUT_INT_RAW_R

Bit 18 - Stores the raw interrupt triggered when touch sensor times out.

pub fn glitch_det_int_raw(&self) -> GLITCH_DET_INT_RAW_R

Bit 19 - Stores the raw interrupt triggered when a glitch is detected.