Module esp32s2::rtc_cntl::options0

Sets the power options of crystal and PLL clocks, and initiates reset by software

Structs

OPTIONS0_SPEC

Sets the power options of crystal and PLL clocks, and initiates reset by software

Type Aliases

ANALOG_FORCE_ISO_R

Field ANALOG_FORCE_ISO reader -

ANALOG_FORCE_ISO_W

Field ANALOG_FORCE_ISO writer -

ANALOG_FORCE_NOISO_R

Field ANALOG_FORCE_NOISO reader -

ANALOG_FORCE_NOISO_W

Field ANALOG_FORCE_NOISO writer -

BBPLL_FORCE_PD_R

Field BBPLL_FORCE_PD reader - Set this bit to FPD BB_PLL.

BBPLL_FORCE_PD_W

Field BBPLL_FORCE_PD writer - Set this bit to FPD BB_PLL.

BBPLL_FORCE_PU_R

Field BBPLL_FORCE_PU reader - Set this bit to FPU BB_PLL.

BBPLL_FORCE_PU_W

Field BBPLL_FORCE_PU writer - Set this bit to FPU BB_PLL.

BBPLL_I2C_FORCE_PD_R

Field BBPLL_I2C_FORCE_PD reader - Set this bit to FPD BB_PLL _I2C.

BBPLL_I2C_FORCE_PD_W

Field BBPLL_I2C_FORCE_PD writer - Set this bit to FPD BB_PLL _I2C.

BBPLL_I2C_FORCE_PU_R

Field BBPLL_I2C_FORCE_PU reader - Set this bit to FPU BB_PLL _I2C.

BBPLL_I2C_FORCE_PU_W

Field BBPLL_I2C_FORCE_PU writer - Set this bit to FPU BB_PLL _I2C.

BB_I2C_FORCE_PD_R

Field BB_I2C_FORCE_PD reader - Set this bit to FPD BB_I2C.

BB_I2C_FORCE_PD_W

Field BB_I2C_FORCE_PD writer - Set this bit to FPD BB_I2C.

BB_I2C_FORCE_PU_R

Field BB_I2C_FORCE_PU reader - Set this bit to FPU BB_I2C.

BB_I2C_FORCE_PU_W

Field BB_I2C_FORCE_PU writer - Set this bit to FPU BB_I2C.

DG_WRAP_FORCE_NORST_R

Field DG_WRAP_FORCE_NORST reader - Set this bit to disable force reset to digital system in deep-sleep.

DG_WRAP_FORCE_NORST_W

Field DG_WRAP_FORCE_NORST writer - Set this bit to disable force reset to digital system in deep-sleep.

DG_WRAP_FORCE_RST_R

Field DG_WRAP_FORCE_RST reader - Set this bit to force reset the digital system in deep-sleep.

DG_WRAP_FORCE_RST_W

Field DG_WRAP_FORCE_RST writer - Set this bit to force reset the digital system in deep-sleep.

PLL_FORCE_ISO_R

Field PLL_FORCE_ISO reader -

PLL_FORCE_ISO_W

Field PLL_FORCE_ISO writer -

PLL_FORCE_NOISO_R

Field PLL_FORCE_NOISO reader -

PLL_FORCE_NOISO_W

Field PLL_FORCE_NOISO writer -

R

Register OPTIONS0 reader

SW_APPCPU_RST_W

Field SW_APPCPU_RST writer - APP CPU SW reset. (Note, we don’t have APP CPU for ESP32-S2)

SW_PROCPU_RST_W

Field SW_PROCPU_RST writer - Set this bit to reset the CPU by SW.

SW_STALL_APPCPU_C0_R

Field SW_STALL_APPCPU_C0 reader - {reg_sw_stall_appcpu_c1[5:0] , reg_sw_stall_appcpu_c0[1:0]} == 0x86 will stall APP CPU

SW_STALL_APPCPU_C0_W

Field SW_STALL_APPCPU_C0 writer - {reg_sw_stall_appcpu_c1[5:0] , reg_sw_stall_appcpu_c0[1:0]} == 0x86 will stall APP CPU

SW_STALL_PROCPU_C0_R

Field SW_STALL_PROCPU_C0 reader - When RTC_CNTL_REG_SW_STALL_PROCPU_C1 is configured to 0x21, setting this bit to 0x2 stalls the CPU by SW.

SW_STALL_PROCPU_C0_W

Field SW_STALL_PROCPU_C0 writer - When RTC_CNTL_REG_SW_STALL_PROCPU_C1 is configured to 0x21, setting this bit to 0x2 stalls the CPU by SW.

SW_SYS_RST_W

Field SW_SYS_RST writer - Set this bit to reset the system via SW.

W

Register OPTIONS0 writer

XTL_FORCE_ISO_R

Field XTL_FORCE_ISO reader -

XTL_FORCE_ISO_W

Field XTL_FORCE_ISO writer -

XTL_FORCE_NOISO_R

Field XTL_FORCE_NOISO reader -

XTL_FORCE_NOISO_W

Field XTL_FORCE_NOISO writer -

XTL_FORCE_PD_R

Field XTL_FORCE_PD reader - Set this bit to FPD the crystal oscillator.

XTL_FORCE_PD_W

Field XTL_FORCE_PD writer - Set this bit to FPD the crystal oscillator.

XTL_FORCE_PU_R

Field XTL_FORCE_PU reader - Set this bit to FPU the crystal oscillator.

XTL_FORCE_PU_W

Field XTL_FORCE_PU writer - Set this bit to FPU the crystal oscillator.