Type Alias esp32s2::rtc_cntl::options0::R

pub type R = R<OPTIONS0_SPEC>;

Register OPTIONS0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn sw_stall_appcpu_c0(&self) -> SW_STALL_APPCPU_C0_R

Bits 0:1 - {reg_sw_stall_appcpu_c1[5:0] , reg_sw_stall_appcpu_c0[1:0]} == 0x86 will stall APP CPU

pub fn sw_stall_procpu_c0(&self) -> SW_STALL_PROCPU_C0_R

Bits 2:3 - When RTC_CNTL_REG_SW_STALL_PROCPU_C1 is configured to 0x21, setting this bit to 0x2 stalls the CPU by SW.

pub fn bb_i2c_force_pd(&self) -> BB_I2C_FORCE_PD_R

Bit 6 - Set this bit to FPD BB_I2C.

pub fn bb_i2c_force_pu(&self) -> BB_I2C_FORCE_PU_R

Bit 7 - Set this bit to FPU BB_I2C.

pub fn bbpll_i2c_force_pd(&self) -> BBPLL_I2C_FORCE_PD_R

Bit 8 - Set this bit to FPD BB_PLL _I2C.

pub fn bbpll_i2c_force_pu(&self) -> BBPLL_I2C_FORCE_PU_R

Bit 9 - Set this bit to FPU BB_PLL _I2C.

pub fn bbpll_force_pd(&self) -> BBPLL_FORCE_PD_R

Bit 10 - Set this bit to FPD BB_PLL.

pub fn bbpll_force_pu(&self) -> BBPLL_FORCE_PU_R

Bit 11 - Set this bit to FPU BB_PLL.

pub fn xtl_force_pd(&self) -> XTL_FORCE_PD_R

Bit 12 - Set this bit to FPD the crystal oscillator.

pub fn xtl_force_pu(&self) -> XTL_FORCE_PU_R

Bit 13 - Set this bit to FPU the crystal oscillator.

pub fn xtl_force_iso(&self) -> XTL_FORCE_ISO_R

Bit 23

pub fn pll_force_iso(&self) -> PLL_FORCE_ISO_R

Bit 24

pub fn analog_force_iso(&self) -> ANALOG_FORCE_ISO_R

Bit 25

pub fn xtl_force_noiso(&self) -> XTL_FORCE_NOISO_R

Bit 26

pub fn pll_force_noiso(&self) -> PLL_FORCE_NOISO_R

Bit 27

pub fn analog_force_noiso(&self) -> ANALOG_FORCE_NOISO_R

Bit 28

pub fn dg_wrap_force_rst(&self) -> DG_WRAP_FORCE_RST_R

Bit 29 - Set this bit to force reset the digital system in deep-sleep.

pub fn dg_wrap_force_norst(&self) -> DG_WRAP_FORCE_NORST_R

Bit 30 - Set this bit to disable force reset to digital system in deep-sleep.