Type Alias esp32s2::rtc_cntl::options0::W

pub type W = W<OPTIONS0_SPEC>;

Register OPTIONS0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn sw_stall_appcpu_c0(&mut self) -> SW_STALL_APPCPU_C0_W<'_, OPTIONS0_SPEC>

Bits 0:1 - {reg_sw_stall_appcpu_c1[5:0] , reg_sw_stall_appcpu_c0[1:0]} == 0x86 will stall APP CPU

pub fn sw_stall_procpu_c0(&mut self) -> SW_STALL_PROCPU_C0_W<'_, OPTIONS0_SPEC>

Bits 2:3 - When RTC_CNTL_REG_SW_STALL_PROCPU_C1 is configured to 0x21, setting this bit to 0x2 stalls the CPU by SW.

pub fn sw_appcpu_rst(&mut self) -> SW_APPCPU_RST_W<'_, OPTIONS0_SPEC>

Bit 4 - APP CPU SW reset. (Note, we don’t have APP CPU for ESP32-S2)

pub fn sw_procpu_rst(&mut self) -> SW_PROCPU_RST_W<'_, OPTIONS0_SPEC>

Bit 5 - Set this bit to reset the CPU by SW.

pub fn bb_i2c_force_pd(&mut self) -> BB_I2C_FORCE_PD_W<'_, OPTIONS0_SPEC>

Bit 6 - Set this bit to FPD BB_I2C.

pub fn bb_i2c_force_pu(&mut self) -> BB_I2C_FORCE_PU_W<'_, OPTIONS0_SPEC>

Bit 7 - Set this bit to FPU BB_I2C.

pub fn bbpll_i2c_force_pd(&mut self) -> BBPLL_I2C_FORCE_PD_W<'_, OPTIONS0_SPEC>

Bit 8 - Set this bit to FPD BB_PLL _I2C.

pub fn bbpll_i2c_force_pu(&mut self) -> BBPLL_I2C_FORCE_PU_W<'_, OPTIONS0_SPEC>

Bit 9 - Set this bit to FPU BB_PLL _I2C.

pub fn bbpll_force_pd(&mut self) -> BBPLL_FORCE_PD_W<'_, OPTIONS0_SPEC>

Bit 10 - Set this bit to FPD BB_PLL.

pub fn bbpll_force_pu(&mut self) -> BBPLL_FORCE_PU_W<'_, OPTIONS0_SPEC>

Bit 11 - Set this bit to FPU BB_PLL.

pub fn xtl_force_pd(&mut self) -> XTL_FORCE_PD_W<'_, OPTIONS0_SPEC>

Bit 12 - Set this bit to FPD the crystal oscillator.

pub fn xtl_force_pu(&mut self) -> XTL_FORCE_PU_W<'_, OPTIONS0_SPEC>

Bit 13 - Set this bit to FPU the crystal oscillator.

pub fn xtl_force_iso(&mut self) -> XTL_FORCE_ISO_W<'_, OPTIONS0_SPEC>

Bit 23

pub fn pll_force_iso(&mut self) -> PLL_FORCE_ISO_W<'_, OPTIONS0_SPEC>

Bit 24

pub fn analog_force_iso(&mut self) -> ANALOG_FORCE_ISO_W<'_, OPTIONS0_SPEC>

Bit 25

pub fn xtl_force_noiso(&mut self) -> XTL_FORCE_NOISO_W<'_, OPTIONS0_SPEC>

Bit 26

pub fn pll_force_noiso(&mut self) -> PLL_FORCE_NOISO_W<'_, OPTIONS0_SPEC>

Bit 27

pub fn analog_force_noiso(&mut self) -> ANALOG_FORCE_NOISO_W<'_, OPTIONS0_SPEC>

Bit 28

pub fn dg_wrap_force_rst(&mut self) -> DG_WRAP_FORCE_RST_W<'_, OPTIONS0_SPEC>

Bit 29 - Set this bit to force reset the digital system in deep-sleep.

pub fn dg_wrap_force_norst( &mut self ) -> DG_WRAP_FORCE_NORST_W<'_, OPTIONS0_SPEC>

Bit 30 - Set this bit to disable force reset to digital system in deep-sleep.

pub fn sw_sys_rst(&mut self) -> SW_SYS_RST_W<'_, OPTIONS0_SPEC>

Bit 31 - Set this bit to reset the system via SW.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual