Module esp32s2::rtc_cntl::pwc

RTC power configuraiton register

Structs

PWC_SPEC

RTC power configuraiton register

Type Aliases

FASTMEM_FOLW_CPU_R

Field FASTMEM_FOLW_CPU reader - Set 1 to FPD the RTC fast memory when the CPU is powered down. Set 0 to FPD the RTC fast memory when the RTC main state machine is powered down.

FASTMEM_FOLW_CPU_W

Field FASTMEM_FOLW_CPU writer - Set 1 to FPD the RTC fast memory when the CPU is powered down. Set 0 to FPD the RTC fast memory when the RTC main state machine is powered down.

FASTMEM_FORCE_ISO_R

Field FASTMEM_FORCE_ISO reader - Set this bit to force isolate the RTC fast memory.

FASTMEM_FORCE_ISO_W

Field FASTMEM_FORCE_ISO writer - Set this bit to force isolate the RTC fast memory.

FASTMEM_FORCE_LPD_R

Field FASTMEM_FORCE_LPD reader - Set this bit to force not retain the RTC fast memory.

FASTMEM_FORCE_LPD_W

Field FASTMEM_FORCE_LPD writer - Set this bit to force not retain the RTC fast memory.

FASTMEM_FORCE_LPU_R

Field FASTMEM_FORCE_LPU reader - Set this bit to force retain the RTC fast memory.

FASTMEM_FORCE_LPU_W

Field FASTMEM_FORCE_LPU writer - Set this bit to force retain the RTC fast memory.

FASTMEM_FORCE_NOISO_R

Field FASTMEM_FORCE_NOISO reader - Set this bit to disable the force isolation to the RTC fast memory.

FASTMEM_FORCE_NOISO_W

Field FASTMEM_FORCE_NOISO writer - Set this bit to disable the force isolation to the RTC fast memory.

FASTMEM_FORCE_PD_R

Field FASTMEM_FORCE_PD reader - Set this bit to FPD the RTC fast memory.

FASTMEM_FORCE_PD_W

Field FASTMEM_FORCE_PD writer - Set this bit to FPD the RTC fast memory.

FASTMEM_FORCE_PU_R

Field FASTMEM_FORCE_PU reader - Set this bit to FPU the RTC fast memory.

FASTMEM_FORCE_PU_W

Field FASTMEM_FORCE_PU writer - Set this bit to FPU the RTC fast memory.

FASTMEM_PD_EN_R

Field FASTMEM_PD_EN reader - Set this bit to enable PD for the RTC fast memory in sleep.

FASTMEM_PD_EN_W

Field FASTMEM_PD_EN writer - Set this bit to enable PD for the RTC fast memory in sleep.

FORCE_ISO_R

Field FORCE_ISO reader - Set this bit to force isolate the RTC peripherals.

FORCE_ISO_W

Field FORCE_ISO writer - Set this bit to force isolate the RTC peripherals.

FORCE_NOISO_R

Field FORCE_NOISO reader - Set this bit to disable the force isolation to the RTC peripherals.

FORCE_NOISO_W

Field FORCE_NOISO writer - Set this bit to disable the force isolation to the RTC peripherals.

FORCE_PD_R

Field FORCE_PD reader - Set this bit to FPD the RTC peripherals.

FORCE_PD_W

Field FORCE_PD writer - Set this bit to FPD the RTC peripherals.

FORCE_PU_R

Field FORCE_PU reader - Set this bit to FPU the RTC peripherals.

FORCE_PU_W

Field FORCE_PU writer - Set this bit to FPU the RTC peripherals.

PAD_FORCE_HOLD_R

Field PAD_FORCE_HOLD reader - Set this bit the force hold the RTC GPIOs.

PAD_FORCE_HOLD_W

Field PAD_FORCE_HOLD writer - Set this bit the force hold the RTC GPIOs.

PD_EN_R

Field PD_EN reader - Set this bit to enable PD for the RTC peripherals in sleep.

PD_EN_W

Field PD_EN writer - Set this bit to enable PD for the RTC peripherals in sleep.

R

Register PWC reader

SLOWMEM_FOLW_CPU_R

Field SLOWMEM_FOLW_CPU reader - Set 1 to FPD the RTC slow memory when the CPU is powered down. Set 0 to FPD the RTC slow memory when the RTC main state machine is powered down.

SLOWMEM_FOLW_CPU_W

Field SLOWMEM_FOLW_CPU writer - Set 1 to FPD the RTC slow memory when the CPU is powered down. Set 0 to FPD the RTC slow memory when the RTC main state machine is powered down.

SLOWMEM_FORCE_ISO_R

Field SLOWMEM_FORCE_ISO reader - Set this bit to force isolate the RTC slow memory.

SLOWMEM_FORCE_ISO_W

Field SLOWMEM_FORCE_ISO writer - Set this bit to force isolate the RTC slow memory.

SLOWMEM_FORCE_LPD_R

Field SLOWMEM_FORCE_LPD reader - Set this bit to force not retain the RTC slow memory.

SLOWMEM_FORCE_LPD_W

Field SLOWMEM_FORCE_LPD writer - Set this bit to force not retain the RTC slow memory.

SLOWMEM_FORCE_LPU_R

Field SLOWMEM_FORCE_LPU reader - Set this bit to force retain the RTC slow memory.

SLOWMEM_FORCE_LPU_W

Field SLOWMEM_FORCE_LPU writer - Set this bit to force retain the RTC slow memory.

SLOWMEM_FORCE_NOISO_R

Field SLOWMEM_FORCE_NOISO reader - Set this bit to disable the force isolation to the RTC slow memory.

SLOWMEM_FORCE_NOISO_W

Field SLOWMEM_FORCE_NOISO writer - Set this bit to disable the force isolation to the RTC slow memory.

SLOWMEM_FORCE_PD_R

Field SLOWMEM_FORCE_PD reader - Set this bit to FPD the RTC slow memory.

SLOWMEM_FORCE_PD_W

Field SLOWMEM_FORCE_PD writer - Set this bit to FPD the RTC slow memory.

SLOWMEM_FORCE_PU_R

Field SLOWMEM_FORCE_PU reader - Set this bit to FPU the RTC slow memory.

SLOWMEM_FORCE_PU_W

Field SLOWMEM_FORCE_PU writer - Set this bit to FPU the RTC slow memory.

SLOWMEM_PD_EN_R

Field SLOWMEM_PD_EN reader - Set this bit to enable PD for the RTC slow memory in sleep.

SLOWMEM_PD_EN_W

Field SLOWMEM_PD_EN writer - Set this bit to enable PD for the RTC slow memory in sleep.

W

Register PWC writer