Module esp32s2::rtc_cntl::reg

RTC/DIG regulator configuration register

Structs

REG_SPEC

RTC/DIG regulator configuration register

Type Aliases

DBIAS_SLP_R

Field DBIAS_SLP reader - Configures the regulation factor for the low-power voltage regulator when the CPU is in sleep status.

DBIAS_SLP_W

Field DBIAS_SLP writer - Configures the regulation factor for the low-power voltage regulator when the CPU is in sleep status.

DBIAS_WAK_R

Field DBIAS_WAK reader - Configures the regulation factor for the low-power voltage regulator when the CPU is in active status.

DBIAS_WAK_W

Field DBIAS_WAK writer - Configures the regulation factor for the low-power voltage regulator when the CPU is in active status.

DBOOST_FORCE_PD_R

Field DBOOST_FORCE_PD reader - RTC_DBOOST force power down

DBOOST_FORCE_PD_W

Field DBOOST_FORCE_PD writer - RTC_DBOOST force power down

DBOOST_FORCE_PU_R

Field DBOOST_FORCE_PU reader - RTC_DBOOST force power up

DBOOST_FORCE_PU_W

Field DBOOST_FORCE_PU writer - RTC_DBOOST force power up

DIG_REG_DBIAS_SLP_R

Field DIG_REG_DBIAS_SLP reader - Configures the regulation factor for the digital system voltage regulator when the CPU is in sleep status.

DIG_REG_DBIAS_SLP_W

Field DIG_REG_DBIAS_SLP writer - Configures the regulation factor for the digital system voltage regulator when the CPU is in sleep status.

DIG_REG_DBIAS_WAK_R

Field DIG_REG_DBIAS_WAK reader - Configures the regulation factor for the digital system voltage regulator when the CPU is in active status.

DIG_REG_DBIAS_WAK_W

Field DIG_REG_DBIAS_WAK writer - Configures the regulation factor for the digital system voltage regulator when the CPU is in active status.

R

Register REG reader

REGULATOR_FORCE_PD_R

Field REGULATOR_FORCE_PD reader - Set this bit to FPD the RTC_REG, which means decreasing its voltage to 0.8 V or lower.

REGULATOR_FORCE_PD_W

Field REGULATOR_FORCE_PD writer - Set this bit to FPD the RTC_REG, which means decreasing its voltage to 0.8 V or lower.

REGULATOR_FORCE_PU_R

Field REGULATOR_FORCE_PU reader - Set this bit to FPU the RTC_REG.

REGULATOR_FORCE_PU_W

Field REGULATOR_FORCE_PU writer - Set this bit to FPU the RTC_REG.

SCK_DCAP_R

Field SCK_DCAP reader - Configures the frequency of the RTC clocks.

SCK_DCAP_W

Field SCK_DCAP writer - Configures the frequency of the RTC clocks.

W

Register REG writer