Module esp32s2::rtc_cntl::touch_ctrl2

Touch control register

Structs

TOUCH_CTRL2_SPEC

Touch control register

Type Aliases

R

Register TOUCH_CTRL2 reader

TOUCH_CLKGATE_EN_R

Field TOUCH_CLKGATE_EN reader - Touch clock enable bit.

TOUCH_CLKGATE_EN_W

Field TOUCH_CLKGATE_EN writer - Touch clock enable bit.

TOUCH_CLK_FO_R

Field TOUCH_CLK_FO reader - Touch clock force on.

TOUCH_CLK_FO_W

Field TOUCH_CLK_FO writer - Touch clock force on.

TOUCH_DBIAS_R

Field TOUCH_DBIAS reader - 0: Use bandgap bias. 1: Use self bias.

TOUCH_DBIAS_W

Field TOUCH_DBIAS writer - 0: Use bandgap bias. 1: Use self bias.

TOUCH_DRANGE_R

Field TOUCH_DRANGE reader - TOUCH attenuation.

TOUCH_DRANGE_W

Field TOUCH_DRANGE writer - TOUCH attenuation.

TOUCH_DREFH_R

Field TOUCH_DREFH reader - TOUCH reference voltage high. 0: 2.4 V 1: 2.5 V 2: 2.6 V 3: 2.7 V.

TOUCH_DREFH_W

Field TOUCH_DREFH writer - TOUCH reference voltage high. 0: 2.4 V 1: 2.5 V 2: 2.6 V 3: 2.7 V.

TOUCH_DREFL_R

Field TOUCH_DREFL reader - TOUCH reference voltage low. 0: 0.5 V 1: 0.6 V 2: 0.7 V 3: 0.8 V.

TOUCH_DREFL_W

Field TOUCH_DREFL writer - TOUCH reference voltage low. 0: 0.5 V 1: 0.6 V 2: 0.7 V 3: 0.8 V.

TOUCH_REFC_R

Field TOUCH_REFC reader - Touch pad 0 reference capacitance.

TOUCH_REFC_W

Field TOUCH_REFC writer - Touch pad 0 reference capacitance.

TOUCH_RESET_R

Field TOUCH_RESET reader - Reset TOUCH FSM via software.

TOUCH_RESET_W

Field TOUCH_RESET writer - Reset TOUCH FSM via software.

TOUCH_SLP_CYC_DIV_R

Field TOUCH_SLP_CYC_DIV reader - When a touch pad is active, sleep cycle could be divided by this number.

TOUCH_SLP_CYC_DIV_W

Field TOUCH_SLP_CYC_DIV writer - When a touch pad is active, sleep cycle could be divided by this number.

TOUCH_SLP_TIMER_EN_R

Field TOUCH_SLP_TIMER_EN reader - Touch timer enable bit.

TOUCH_SLP_TIMER_EN_W

Field TOUCH_SLP_TIMER_EN writer - Touch timer enable bit.

TOUCH_START_EN_R

Field TOUCH_START_EN reader - 1: Start the Touch FSM, only valid when RTC_CNTL_TOUCH_START_FORCE = 1.

TOUCH_START_EN_W

Field TOUCH_START_EN writer - 1: Start the Touch FSM, only valid when RTC_CNTL_TOUCH_START_FORCE = 1.

TOUCH_START_FORCE_R

Field TOUCH_START_FORCE reader - 0: Start the Touch FSM by timer. 1: Start Touch FSM by software.

TOUCH_START_FORCE_W

Field TOUCH_START_FORCE writer - 0: Start the Touch FSM by timer. 1: Start Touch FSM by software.

TOUCH_START_FSM_EN_R

Field TOUCH_START_FSM_EN reader - 0: TOUCH_START and TOUCH_XPD are controlled by soft- ware. 1: TOUCH_START and TOUCH_XPD are controlled by the Touch FSM.

TOUCH_START_FSM_EN_W

Field TOUCH_START_FSM_EN writer - 0: TOUCH_START and TOUCH_XPD are controlled by soft- ware. 1: TOUCH_START and TOUCH_XPD are controlled by the Touch FSM.

TOUCH_TIMER_FORCE_DONE_R

Field TOUCH_TIMER_FORCE_DONE reader - Force touch timer done.

TOUCH_TIMER_FORCE_DONE_W

Field TOUCH_TIMER_FORCE_DONE writer - Force touch timer done.

TOUCH_XPD_BIAS_R

Field TOUCH_XPD_BIAS reader - TOUCH BIAS power switch.

TOUCH_XPD_BIAS_W

Field TOUCH_XPD_BIAS writer - TOUCH BIAS power switch.

TOUCH_XPD_WAIT_R

Field TOUCH_XPD_WAIT reader - The waiting cycles (in 8MHz) between TOUCH_START and TOUCH_XPD.

TOUCH_XPD_WAIT_W

Field TOUCH_XPD_WAIT writer - The waiting cycles (in 8MHz) between TOUCH_START and TOUCH_XPD.

W

Register TOUCH_CTRL2 writer