Module esp32s2::rtc_i2c::ctrl

Transmission setting

Structs

CTRL_SPEC

Transmission setting

Type Aliases

CLK_EN_R

Field CLK_EN reader - rtc i2c reg clk gating

CLK_EN_W

Field CLK_EN writer - rtc i2c reg clk gating

CLK_GATE_EN_R

Field CLK_GATE_EN reader - RTC I²C controller clock gate.

CLK_GATE_EN_W

Field CLK_GATE_EN writer - RTC I²C controller clock gate.

MS_MODE_R

Field MS_MODE reader - Set this bit to configure RTC I²C as a master.

MS_MODE_W

Field MS_MODE writer - Set this bit to configure RTC I²C as a master.

R

Register CTRL reader

RESET_R

Field RESET reader - RTC I²C software reset.

RESET_W

Field RESET writer - RTC I²C software reset.

RX_LSB_FIRST_R

Field RX_LSB_FIRST reader - This bit is used to control the storage mode for received data. 0: receive data from the most significant bit. 1: receive data from the least significant bit.

RX_LSB_FIRST_W

Field RX_LSB_FIRST writer - This bit is used to control the storage mode for received data. 0: receive data from the most significant bit. 1: receive data from the least significant bit.

SCL_FORCE_OUT_R

Field SCL_FORCE_OUT reader - SCL output mode. 0: open drain. 1: push pull.

SCL_FORCE_OUT_W

Field SCL_FORCE_OUT writer - SCL output mode. 0: open drain. 1: push pull.

SDA_FORCE_OUT_R

Field SDA_FORCE_OUT reader - SDA output mode. 0: open drain. 1: push pull.

SDA_FORCE_OUT_W

Field SDA_FORCE_OUT writer - SDA output mode. 0: open drain. 1: push pull.

TRANS_START_R

Field TRANS_START reader - Set this bit to 1, RTC I2C starts sending data.

TRANS_START_W

Field TRANS_START writer - Set this bit to 1, RTC I2C starts sending data.

TX_LSB_FIRST_R

Field TX_LSB_FIRST reader - This bit is used to control the sending mode. 0: send data from the most significant bit. 1: send data from the least significant bit.

TX_LSB_FIRST_W

Field TX_LSB_FIRST writer - This bit is used to control the sending mode. 0: send data from the most significant bit. 1: send data from the least significant bit.

W

Register CTRL writer