Module esp32s2::rtc_io::xtl_ext_ctr

Crystal power down enable GPIO source

Structs

XTL_EXT_CTR_SPEC

Crystal power down enable GPIO source

Type Aliases

R

Register XTL_EXT_CTR reader

SEL_R

Field SEL reader - Select the external crystal power down enable source to get into sleep mode. 0: select GPIO0. 1: select GPIO1, etc. The input value on this pin XOR RTC_CNTL_EXT_XTL_CONF_REG[30] is the crystal power down enable signal.

SEL_W

Field SEL writer - Select the external crystal power down enable source to get into sleep mode. 0: select GPIO0. 1: select GPIO1, etc. The input value on this pin XOR RTC_CNTL_EXT_XTL_CONF_REG[30] is the crystal power down enable signal.

W

Register XTL_EXT_CTR writer