Module esp32s2::sha

SHA (Secure Hash Algorithm) Accelerator

Modules

busy

Indicates if SHA Accelerator is busy or not

continue_

Continues SHA operation (only effective in Typical SHA mode)

date

Version control register

dma_block_num

Block number register (only effective for DMA-SHA)

dma_continue

Continues SHA operation (only effective in DMA-SHA mode)

dma_start

Starts the SHA accelerator for DMA-SHA operation

h_mem

Hash value

int_clear

DMA-SHA interrupt clear register

int_ena

DMA-SHA interrupt enable register

m_mem

Message

mode

Defines the algorithm of SHA accelerator

start

Starts the SHA accelerator for Typical SHA operation

t_length

String length register for calculating initial Hash Value (only effective for SHA-512/t)

t_string

String content register for calculating initial Hash Value (only effective for SHA-512/t)

Structs

RegisterBlock

Register block

Type Aliases

BUSY

BUSY (r) register accessor: Indicates if SHA Accelerator is busy or not

CONTINUE

CONTINUE (w) register accessor: Continues SHA operation (only effective in Typical SHA mode)

DATE

DATE (rw) register accessor: Version control register

DMA_BLOCK_NUM

DMA_BLOCK_NUM (rw) register accessor: Block number register (only effective for DMA-SHA)

DMA_CONTINUE

DMA_CONTINUE (w) register accessor: Continues SHA operation (only effective in DMA-SHA mode)

DMA_START

DMA_START (w) register accessor: Starts the SHA accelerator for DMA-SHA operation

H_MEM

H_MEM (rw) register accessor: Hash value

INT_CLEAR

INT_CLEAR (w) register accessor: DMA-SHA interrupt clear register

INT_ENA

INT_ENA (rw) register accessor: DMA-SHA interrupt enable register

MODE

MODE (rw) register accessor: Defines the algorithm of SHA accelerator

M_MEM

M_MEM (rw) register accessor: Message

START

START (w) register accessor: Starts the SHA accelerator for Typical SHA operation

T_LENGTH

T_LENGTH (rw) register accessor: String length register for calculating initial Hash Value (only effective for SHA-512/t)

T_STRING

T_STRING (rw) register accessor: String content register for calculating initial Hash Value (only effective for SHA-512/t)