Module esp32s2::spi0::dma_conf

SPI DMA control register

Structs

DMA_CONF_SPEC

SPI DMA control register

Type Aliases

AHBM_FIFO_RST_R

Field AHBM_FIFO_RST reader - Reset spi dma ahb master fifo pointer.

AHBM_FIFO_RST_W

Field AHBM_FIFO_RST writer - Reset spi dma ahb master fifo pointer.

AHBM_RST_R

Field AHBM_RST reader - Reset spi dma ahb master.

AHBM_RST_W

Field AHBM_RST writer - Reset spi dma ahb master.

DMA_CONTINUE_R

Field DMA_CONTINUE reader - spi dma continue tx/rx data.

DMA_CONTINUE_W

Field DMA_CONTINUE writer - spi dma continue tx/rx data.

DMA_INFIFO_FULL_CLR_R

Field DMA_INFIFO_FULL_CLR reader - 1:Clear spi_dma_infifo_full_vld. 0: Do not control it.

DMA_INFIFO_FULL_CLR_W

Field DMA_INFIFO_FULL_CLR writer - 1:Clear spi_dma_infifo_full_vld. 0: Do not control it.

DMA_OUTFIFO_EMPTY_CLR_R

Field DMA_OUTFIFO_EMPTY_CLR reader - 1:Clear spi_dma_outfifo_empty_vld. 0: Do not control it.

DMA_OUTFIFO_EMPTY_CLR_W

Field DMA_OUTFIFO_EMPTY_CLR writer - 1:Clear spi_dma_outfifo_empty_vld. 0: Do not control it.

DMA_RX_STOP_R

Field DMA_RX_STOP reader - spi dma read data stop when in continue tx/rx mode.

DMA_RX_STOP_W

Field DMA_RX_STOP writer - spi dma read data stop when in continue tx/rx mode.

DMA_SEG_TRANS_CLR_R

Field DMA_SEG_TRANS_CLR reader - 1: End slave seg-trans, which acts as 0x05 command. 2 or more end seg-trans signals will induce error in DMA RX. 0: others. Will be cleared in 1 APB CLK cycles by hardware..

DMA_SEG_TRANS_CLR_W

Field DMA_SEG_TRANS_CLR writer - 1: End slave seg-trans, which acts as 0x05 command. 2 or more end seg-trans signals will induce error in DMA RX. 0: others. Will be cleared in 1 APB CLK cycles by hardware..

DMA_SLV_SEG_TRANS_EN_R

Field DMA_SLV_SEG_TRANS_EN reader - Enable dma segment transfer in spi dma half slave mode. 1: enable. 0: disable.

DMA_SLV_SEG_TRANS_EN_W

Field DMA_SLV_SEG_TRANS_EN writer - Enable dma segment transfer in spi dma half slave mode. 1: enable. 0: disable.

DMA_TX_STOP_R

Field DMA_TX_STOP reader - spi dma write data stop when in continue tx/rx mode.

DMA_TX_STOP_W

Field DMA_TX_STOP writer - spi dma write data stop when in continue tx/rx mode.

EXT_MEM_BK_SIZE_R

Field EXT_MEM_BK_SIZE reader - Select the external memory block size.

EXT_MEM_BK_SIZE_W

Field EXT_MEM_BK_SIZE writer - Select the external memory block size.

INDSCR_BURST_EN_R

Field INDSCR_BURST_EN reader - read descriptor use burst mode when write data to memory.

INDSCR_BURST_EN_W

Field INDSCR_BURST_EN writer - read descriptor use burst mode when write data to memory.

IN_LOOP_TEST_R

Field IN_LOOP_TEST reader - Set bit to test in link.

IN_LOOP_TEST_W

Field IN_LOOP_TEST writer - Set bit to test in link.

IN_RST_R

Field IN_RST reader - The bit is used to reset in dma fsm and in data fifo pointer.

IN_RST_W

Field IN_RST writer - The bit is used to reset in dma fsm and in data fifo pointer.

MEM_TRANS_EN_R

Field MEM_TRANS_EN reader - 1: Internal memory data transfer enable bit. Send SPI DMA RX buffer data to SPI DMA TX buffer. 0: Disable this function.

MEM_TRANS_EN_W

Field MEM_TRANS_EN writer - 1: Internal memory data transfer enable bit. Send SPI DMA RX buffer data to SPI DMA TX buffer. 0: Disable this function.

OUTDSCR_BURST_EN_R

Field OUTDSCR_BURST_EN reader - read descriptor use burst mode when read data for memory.

OUTDSCR_BURST_EN_W

Field OUTDSCR_BURST_EN writer - read descriptor use burst mode when read data for memory.

OUT_AUTO_WRBACK_R

Field OUT_AUTO_WRBACK reader - when the bit is set, DMA continue to use the next inlink node when the length of inlink is 0.

OUT_AUTO_WRBACK_W

Field OUT_AUTO_WRBACK writer - when the bit is set, DMA continue to use the next inlink node when the length of inlink is 0.

OUT_DATA_BURST_EN_R

Field OUT_DATA_BURST_EN reader - spi dma read data from memory in burst mode.

OUT_DATA_BURST_EN_W

Field OUT_DATA_BURST_EN writer - spi dma read data from memory in burst mode.

OUT_EOF_MODE_R

Field OUT_EOF_MODE reader - out eof flag generation mode . 1: when dma pop all data from fifo 0:when ahb push all data to fifo.

OUT_EOF_MODE_W

Field OUT_EOF_MODE writer - out eof flag generation mode . 1: when dma pop all data from fifo 0:when ahb push all data to fifo.

OUT_LOOP_TEST_R

Field OUT_LOOP_TEST reader - Set bit to test out link.

OUT_LOOP_TEST_W

Field OUT_LOOP_TEST writer - Set bit to test out link.

OUT_RST_R

Field OUT_RST reader - The bit is used to reset out dma fsm and out data fifo pointer.

OUT_RST_W

Field OUT_RST writer - The bit is used to reset out dma fsm and out data fifo pointer.

R

Register DMA_CONF reader

RX_EOF_EN_R

Field RX_EOF_EN reader - 1: SPI_IN_SUC_EOF_INT_RAW is set when the number of dma pushed data bytes is equal to the value of SPI_SLV_DMA_RD_BYTELEN[19:0]/ SPI_MST_DMA_RD_BYTELEN[19:0] in spi dma transition. 0: SPI_IN_SUC_EOF_INT_RAW is set by SPI_TRANS_DONE in non-seg-trans or SPI_DMA_SEG_TRANS_DONE in seg-trans.

RX_EOF_EN_W

Field RX_EOF_EN writer - 1: SPI_IN_SUC_EOF_INT_RAW is set when the number of dma pushed data bytes is equal to the value of SPI_SLV_DMA_RD_BYTELEN[19:0]/ SPI_MST_DMA_RD_BYTELEN[19:0] in spi dma transition. 0: SPI_IN_SUC_EOF_INT_RAW is set by SPI_TRANS_DONE in non-seg-trans or SPI_DMA_SEG_TRANS_DONE in seg-trans.

SLV_LAST_SEG_POP_CLR_R

Field SLV_LAST_SEG_POP_CLR reader - 1: Clear spi_slv_seg_frt_pop_mask. 0 : others

SLV_LAST_SEG_POP_CLR_W

Field SLV_LAST_SEG_POP_CLR writer - 1: Clear spi_slv_seg_frt_pop_mask. 0 : others

SLV_RX_SEG_TRANS_CLR_EN_R

Field SLV_RX_SEG_TRANS_CLR_EN reader - 1: spi_dma_infifo_full_vld is cleared by spi slave CMD5. 0: spi_dma_infifo_full_vld is cleared by SPI_TRANS_DONE.

SLV_RX_SEG_TRANS_CLR_EN_W

Field SLV_RX_SEG_TRANS_CLR_EN writer - 1: spi_dma_infifo_full_vld is cleared by spi slave CMD5. 0: spi_dma_infifo_full_vld is cleared by SPI_TRANS_DONE.

SLV_TX_SEG_TRANS_CLR_EN_R

Field SLV_TX_SEG_TRANS_CLR_EN reader - 1: spi_dma_outfifo_empty_vld is cleared by spi slave CMD6. 0: spi_dma_outfifo_empty_vld is cleared by SPI_TRANS_DONE.

SLV_TX_SEG_TRANS_CLR_EN_W

Field SLV_TX_SEG_TRANS_CLR_EN writer - 1: spi_dma_outfifo_empty_vld is cleared by spi slave CMD6. 0: spi_dma_outfifo_empty_vld is cleared by SPI_TRANS_DONE.

W

Register DMA_CONF writer