Type Alias esp32s2::spi0::dma_conf::W

pub type W = W<DMA_CONF_SPEC>;

Register DMA_CONF writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn in_rst(&mut self) -> IN_RST_W<'_, DMA_CONF_SPEC>

Bit 2 - The bit is used to reset in dma fsm and in data fifo pointer.

pub fn out_rst(&mut self) -> OUT_RST_W<'_, DMA_CONF_SPEC>

Bit 3 - The bit is used to reset out dma fsm and out data fifo pointer.

pub fn ahbm_fifo_rst(&mut self) -> AHBM_FIFO_RST_W<'_, DMA_CONF_SPEC>

Bit 4 - Reset spi dma ahb master fifo pointer.

pub fn ahbm_rst(&mut self) -> AHBM_RST_W<'_, DMA_CONF_SPEC>

Bit 5 - Reset spi dma ahb master.

pub fn in_loop_test(&mut self) -> IN_LOOP_TEST_W<'_, DMA_CONF_SPEC>

Bit 6 - Set bit to test in link.

pub fn out_loop_test(&mut self) -> OUT_LOOP_TEST_W<'_, DMA_CONF_SPEC>

Bit 7 - Set bit to test out link.

pub fn out_auto_wrback(&mut self) -> OUT_AUTO_WRBACK_W<'_, DMA_CONF_SPEC>

Bit 8 - when the bit is set, DMA continue to use the next inlink node when the length of inlink is 0.

pub fn out_eof_mode(&mut self) -> OUT_EOF_MODE_W<'_, DMA_CONF_SPEC>

Bit 9 - out eof flag generation mode . 1: when dma pop all data from fifo 0:when ahb push all data to fifo.

pub fn outdscr_burst_en(&mut self) -> OUTDSCR_BURST_EN_W<'_, DMA_CONF_SPEC>

Bit 10 - read descriptor use burst mode when read data for memory.

pub fn indscr_burst_en(&mut self) -> INDSCR_BURST_EN_W<'_, DMA_CONF_SPEC>

Bit 11 - read descriptor use burst mode when write data to memory.

pub fn out_data_burst_en(&mut self) -> OUT_DATA_BURST_EN_W<'_, DMA_CONF_SPEC>

Bit 12 - spi dma read data from memory in burst mode.

pub fn mem_trans_en(&mut self) -> MEM_TRANS_EN_W<'_, DMA_CONF_SPEC>

Bit 13 - 1: Internal memory data transfer enable bit. Send SPI DMA RX buffer data to SPI DMA TX buffer. 0: Disable this function.

pub fn dma_rx_stop(&mut self) -> DMA_RX_STOP_W<'_, DMA_CONF_SPEC>

Bit 14 - spi dma read data stop when in continue tx/rx mode.

pub fn dma_tx_stop(&mut self) -> DMA_TX_STOP_W<'_, DMA_CONF_SPEC>

Bit 15 - spi dma write data stop when in continue tx/rx mode.

pub fn dma_continue(&mut self) -> DMA_CONTINUE_W<'_, DMA_CONF_SPEC>

Bit 16 - spi dma continue tx/rx data.

pub fn slv_last_seg_pop_clr( &mut self ) -> SLV_LAST_SEG_POP_CLR_W<'_, DMA_CONF_SPEC>

Bit 17 - 1: Clear spi_slv_seg_frt_pop_mask. 0 : others

pub fn dma_slv_seg_trans_en( &mut self ) -> DMA_SLV_SEG_TRANS_EN_W<'_, DMA_CONF_SPEC>

Bit 18 - Enable dma segment transfer in spi dma half slave mode. 1: enable. 0: disable.

pub fn slv_rx_seg_trans_clr_en( &mut self ) -> SLV_RX_SEG_TRANS_CLR_EN_W<'_, DMA_CONF_SPEC>

Bit 19 - 1: spi_dma_infifo_full_vld is cleared by spi slave CMD5. 0: spi_dma_infifo_full_vld is cleared by SPI_TRANS_DONE.

pub fn slv_tx_seg_trans_clr_en( &mut self ) -> SLV_TX_SEG_TRANS_CLR_EN_W<'_, DMA_CONF_SPEC>

Bit 20 - 1: spi_dma_outfifo_empty_vld is cleared by spi slave CMD6. 0: spi_dma_outfifo_empty_vld is cleared by SPI_TRANS_DONE.

pub fn rx_eof_en(&mut self) -> RX_EOF_EN_W<'_, DMA_CONF_SPEC>

Bit 21 - 1: SPI_IN_SUC_EOF_INT_RAW is set when the number of dma pushed data bytes is equal to the value of SPI_SLV_DMA_RD_BYTELEN[19:0]/ SPI_MST_DMA_RD_BYTELEN[19:0] in spi dma transition. 0: SPI_IN_SUC_EOF_INT_RAW is set by SPI_TRANS_DONE in non-seg-trans or SPI_DMA_SEG_TRANS_DONE in seg-trans.

pub fn dma_infifo_full_clr( &mut self ) -> DMA_INFIFO_FULL_CLR_W<'_, DMA_CONF_SPEC>

Bit 22 - 1:Clear spi_dma_infifo_full_vld. 0: Do not control it.

pub fn dma_outfifo_empty_clr( &mut self ) -> DMA_OUTFIFO_EMPTY_CLR_W<'_, DMA_CONF_SPEC>

Bit 23 - 1:Clear spi_dma_outfifo_empty_vld. 0: Do not control it.

pub fn ext_mem_bk_size(&mut self) -> EXT_MEM_BK_SIZE_W<'_, DMA_CONF_SPEC>

Bits 26:27 - Select the external memory block size.

pub fn dma_seg_trans_clr(&mut self) -> DMA_SEG_TRANS_CLR_W<'_, DMA_CONF_SPEC>

Bit 28 - 1: End slave seg-trans, which acts as 0x05 command. 2 or more end seg-trans signals will induce error in DMA RX. 0: others. Will be cleared in 1 APB CLK cycles by hardware..

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual