Type Alias esp32s2::spi0::dma_int_clr::R

pub type R = R<DMA_INT_CLR_SPEC>;

Register DMA_INT_CLR reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn inlink_dscr_empty_int_clr(&self) -> INLINK_DSCR_EMPTY_INT_CLR_R

Bit 0 - The clear bit for lack of enough inlink descriptors. Can be configured in CONF state.

pub fn outlink_dscr_error_int_clr(&self) -> OUTLINK_DSCR_ERROR_INT_CLR_R

Bit 1 - The clear bit for outlink descriptor error. Can be configured in CONF state.

pub fn inlink_dscr_error_int_clr(&self) -> INLINK_DSCR_ERROR_INT_CLR_R

Bit 2 - The clear bit for inlink descriptor error. Can be configured in CONF state.

pub fn in_done_int_clr(&self) -> IN_DONE_INT_CLR_R

Bit 3 - The clear bit for completing usage of a inlink descriptor. Can be configured in CONF state.

pub fn in_err_eof_int_clr(&self) -> IN_ERR_EOF_INT_CLR_R

Bit 4 - The clear bit for receiving error. Can be configured in CONF state.

pub fn in_suc_eof_int_clr(&self) -> IN_SUC_EOF_INT_CLR_R

Bit 5 - The clear bit for completing receiving all the packets from host. Can be configured in CONF state.

pub fn out_done_int_clr(&self) -> OUT_DONE_INT_CLR_R

Bit 6 - The clear bit for completing usage of a outlink descriptor. Can be configured in CONF state.

pub fn out_eof_int_clr(&self) -> OUT_EOF_INT_CLR_R

Bit 7 - The clear bit for sending a packet to host done. Can be configured in CONF state.

pub fn out_total_eof_int_clr(&self) -> OUT_TOTAL_EOF_INT_CLR_R

Bit 8 - The clear bit for sending all the packets to host done. Can be configured in CONF state.

pub fn infifo_full_err_int_clr(&self) -> INFIFO_FULL_ERR_INT_CLR_R

Bit 9 - 1: Clear SPI_INFIFO_FULL_ERR_INT_RAW. 0: not valid. Can be changed by CONF_buf.

pub fn outfifo_empty_err_int_clr(&self) -> OUTFIFO_EMPTY_ERR_INT_CLR_R

Bit 10 - 1: Clear SPI_OUTFIFO_EMPTY_ERR_INT_RAW signal. 0: not valid. Can be changed by CONF_buf.

pub fn slv_cmd6_int_clr(&self) -> SLV_CMD6_INT_CLR_R

Bit 11 - The clear bit for SPI slave CMD6 interrupt.

pub fn slv_cmd7_int_clr(&self) -> SLV_CMD7_INT_CLR_R

Bit 12 - The clear bit for SPI slave CMD7 interrupt.

pub fn slv_cmd8_int_clr(&self) -> SLV_CMD8_INT_CLR_R

Bit 13 - The clear bit for SPI slave CMD8 interrupt.

pub fn slv_cmd9_int_clr(&self) -> SLV_CMD9_INT_CLR_R

Bit 14 - The clear bit for SPI slave CMD9 interrupt.

pub fn slv_cmda_int_clr(&self) -> SLV_CMDA_INT_CLR_R

Bit 15 - The clear bit for SPI slave CMDA interrupt.