Type Alias esp32s2::spi0::dma_int_ena::W

pub type W = W<DMA_INT_ENA_SPEC>;

Register DMA_INT_ENA writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn inlink_dscr_empty_int_ena( &mut self ) -> INLINK_DSCR_EMPTY_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 0 - The enable bit for lack of enough inlink descriptors. Can be configured in CONF state.

pub fn outlink_dscr_error_int_ena( &mut self ) -> OUTLINK_DSCR_ERROR_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 1 - The enable bit for outlink descriptor error. Can be configured in CONF state.

pub fn inlink_dscr_error_int_ena( &mut self ) -> INLINK_DSCR_ERROR_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 2 - The enable bit for inlink descriptor error. Can be configured in CONF state.

pub fn in_done_int_ena(&mut self) -> IN_DONE_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 3 - The enable bit for completing usage of a inlink descriptor. Can be configured in CONF state.

pub fn in_err_eof_int_ena( &mut self ) -> IN_ERR_EOF_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 4 - The enable bit for receiving error. Can be configured in CONF state.

pub fn in_suc_eof_int_ena( &mut self ) -> IN_SUC_EOF_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 5 - The enable bit for completing receiving all the packets from host. Can be configured in CONF state.

pub fn out_done_int_ena(&mut self) -> OUT_DONE_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 6 - The enable bit for completing usage of a outlink descriptor . Can be configured in CONF state.

pub fn out_eof_int_ena(&mut self) -> OUT_EOF_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 7 - The enable bit for sending a packet to host done. Can be configured in CONF state.

pub fn out_total_eof_int_ena( &mut self ) -> OUT_TOTAL_EOF_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 8 - The enable bit for sending all the packets to host done. Can be configured in CONF state.

pub fn infifo_full_err_int_ena( &mut self ) -> INFIFO_FULL_ERR_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 9 - The enable bit for infifo full error interrupt.

pub fn outfifo_empty_err_int_ena( &mut self ) -> OUTFIFO_EMPTY_ERR_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 10 - The enable bit for outfifo empty error interrupt.

pub fn slv_cmd6_int_ena(&mut self) -> SLV_CMD6_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 11 - The enable bit for SPI slave CMD6 interrupt.

pub fn slv_cmd7_int_ena(&mut self) -> SLV_CMD7_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 12 - The enable bit for SPI slave CMD7 interrupt.

pub fn slv_cmd8_int_ena(&mut self) -> SLV_CMD8_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 13 - The enable bit for SPI slave CMD8 interrupt.

pub fn slv_cmd9_int_ena(&mut self) -> SLV_CMD9_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 14 - The enable bit for SPI slave CMD9 interrupt.

pub fn slv_cmda_int_ena(&mut self) -> SLV_CMDA_INT_ENA_W<'_, DMA_INT_ENA_SPEC>

Bit 15 - The enable bit for SPI slave CMDA interrupt.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual