Module esp32s2::spi0::dma_int_st

SPI DMA interrupt status register

Structs

DMA_INT_ST_SPEC

SPI DMA interrupt status register

Type Aliases

INFIFO_FULL_ERR_INT_ST_R

Field INFIFO_FULL_ERR_INT_ST reader - The status bit for infifo full error.

INLINK_DSCR_EMPTY_INT_ST_R

Field INLINK_DSCR_EMPTY_INT_ST reader - The status bit for lack of enough inlink descriptors.

INLINK_DSCR_ERROR_INT_ST_R

Field INLINK_DSCR_ERROR_INT_ST reader - The status bit for inlink descriptor error.

IN_DONE_INT_ST_R

Field IN_DONE_INT_ST reader - The status bit for completing usage of a inlink descriptor.

IN_ERR_EOF_INT_ST_R

Field IN_ERR_EOF_INT_ST reader - The status bit for receiving error.

IN_SUC_EOF_INT_ST_R

Field IN_SUC_EOF_INT_ST reader - The status bit for completing receiving all the packets from host.

OUTFIFO_EMPTY_ERR_INT_ST_R

Field OUTFIFO_EMPTY_ERR_INT_ST reader - The status bit for outfifo empty error.

OUTLINK_DSCR_ERROR_INT_ST_R

Field OUTLINK_DSCR_ERROR_INT_ST reader - The status bit for outlink descriptor error.

OUT_DONE_INT_ST_R

Field OUT_DONE_INT_ST reader - The status bit for completing usage of a outlink descriptor.

OUT_EOF_INT_ST_R

Field OUT_EOF_INT_ST reader - The status bit for sending a packet to host done.

OUT_TOTAL_EOF_INT_ST_R

Field OUT_TOTAL_EOF_INT_ST reader - The status bit for sending all the packets to host done.

R

Register DMA_INT_ST reader

SLV_CMD6_INT_ST_R

Field SLV_CMD6_INT_ST reader - The status bit for SPI slave CMD6 interrupt.

SLV_CMD6_INT_ST_W

Field SLV_CMD6_INT_ST writer - The status bit for SPI slave CMD6 interrupt.

SLV_CMD7_INT_ST_R

Field SLV_CMD7_INT_ST reader - The status bit for SPI slave CMD7 interrupt.

SLV_CMD7_INT_ST_W

Field SLV_CMD7_INT_ST writer - The status bit for SPI slave CMD7 interrupt.

SLV_CMD8_INT_ST_R

Field SLV_CMD8_INT_ST reader - The status bit for SPI slave CMD8 interrupt.

SLV_CMD8_INT_ST_W

Field SLV_CMD8_INT_ST writer - The status bit for SPI slave CMD8 interrupt.

SLV_CMD9_INT_ST_R

Field SLV_CMD9_INT_ST reader - The status bit for SPI slave CMD9 interrupt.

SLV_CMD9_INT_ST_W

Field SLV_CMD9_INT_ST writer - The status bit for SPI slave CMD9 interrupt.

SLV_CMDA_INT_ST_R

Field SLV_CMDA_INT_ST reader - The status bit for SPI slave CMDA interrupt.

SLV_CMDA_INT_ST_W

Field SLV_CMDA_INT_ST writer - The status bit for SPI slave CMDA interrupt.

W

Register DMA_INT_ST writer