Type Alias esp32s2::spi0::dma_int_st::R

pub type R = R<DMA_INT_ST_SPEC>;

Register DMA_INT_ST reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn inlink_dscr_empty_int_st(&self) -> INLINK_DSCR_EMPTY_INT_ST_R

Bit 0 - The status bit for lack of enough inlink descriptors.

pub fn outlink_dscr_error_int_st(&self) -> OUTLINK_DSCR_ERROR_INT_ST_R

Bit 1 - The status bit for outlink descriptor error.

pub fn inlink_dscr_error_int_st(&self) -> INLINK_DSCR_ERROR_INT_ST_R

Bit 2 - The status bit for inlink descriptor error.

pub fn in_done_int_st(&self) -> IN_DONE_INT_ST_R

Bit 3 - The status bit for completing usage of a inlink descriptor.

pub fn in_err_eof_int_st(&self) -> IN_ERR_EOF_INT_ST_R

Bit 4 - The status bit for receiving error.

pub fn in_suc_eof_int_st(&self) -> IN_SUC_EOF_INT_ST_R

Bit 5 - The status bit for completing receiving all the packets from host.

pub fn out_done_int_st(&self) -> OUT_DONE_INT_ST_R

Bit 6 - The status bit for completing usage of a outlink descriptor.

pub fn out_eof_int_st(&self) -> OUT_EOF_INT_ST_R

Bit 7 - The status bit for sending a packet to host done.

pub fn out_total_eof_int_st(&self) -> OUT_TOTAL_EOF_INT_ST_R

Bit 8 - The status bit for sending all the packets to host done.

pub fn infifo_full_err_int_st(&self) -> INFIFO_FULL_ERR_INT_ST_R

Bit 9 - The status bit for infifo full error.

pub fn outfifo_empty_err_int_st(&self) -> OUTFIFO_EMPTY_ERR_INT_ST_R

Bit 10 - The status bit for outfifo empty error.

pub fn slv_cmd6_int_st(&self) -> SLV_CMD6_INT_ST_R

Bit 11 - The status bit for SPI slave CMD6 interrupt.

pub fn slv_cmd7_int_st(&self) -> SLV_CMD7_INT_ST_R

Bit 12 - The status bit for SPI slave CMD7 interrupt.

pub fn slv_cmd8_int_st(&self) -> SLV_CMD8_INT_ST_R

Bit 13 - The status bit for SPI slave CMD8 interrupt.

pub fn slv_cmd9_int_st(&self) -> SLV_CMD9_INT_ST_R

Bit 14 - The status bit for SPI slave CMD9 interrupt.

pub fn slv_cmda_int_st(&self) -> SLV_CMDA_INT_ST_R

Bit 15 - The status bit for SPI slave CMDA interrupt.