Module esp32s2::spi0::dout_num

SPI output delay number configuration

Structs

DOUT_NUM_SPEC

SPI output delay number configuration

Type Aliases

DOUT0_NUM_R

Field DOUT0_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT0_NUM_W

Field DOUT0_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT1_NUM_R

Field DOUT1_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT1_NUM_W

Field DOUT1_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT2_NUM_R

Field DOUT2_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT2_NUM_W

Field DOUT2_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT3_NUM_R

Field DOUT3_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT3_NUM_W

Field DOUT3_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT4_NUM_R

Field DOUT4_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT4_NUM_W

Field DOUT4_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT5_NUM_R

Field DOUT5_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT5_NUM_W

Field DOUT5_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT6_NUM_R

Field DOUT6_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT6_NUM_W

Field DOUT6_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT7_NUM_R

Field DOUT7_NUM reader - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

DOUT7_NUM_W

Field DOUT7_NUM writer - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

R

Register DOUT_NUM reader

W

Register DOUT_NUM writer