Type Alias esp32s2::spi0::dout_num::R

pub type R = R<DOUT_NUM_SPEC>;

Register DOUT_NUM reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn dout0_num(&self) -> DOUT0_NUM_R

Bits 0:1 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout1_num(&self) -> DOUT1_NUM_R

Bits 2:3 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout2_num(&self) -> DOUT2_NUM_R

Bits 4:5 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout3_num(&self) -> DOUT3_NUM_R

Bits 6:7 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout4_num(&self) -> DOUT4_NUM_R

Bits 8:9 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout5_num(&self) -> DOUT5_NUM_R

Bits 10:11 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout6_num(&self) -> DOUT6_NUM_R

Bits 12:13 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.

pub fn dout7_num(&self) -> DOUT7_NUM_R

Bits 14:15 - the output signals are delayed by system clock cycles, 0: delayed by 1 cycle, 1: delayed by 2 cycles,… Can be configured in CONF state.