Type Alias esp32s2::spi0::lcd_d_mode::R

pub type R = R<LCD_D_MODE_SPEC>;

Register LCD_D_MODE reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn d_dqs_mode(&self) -> D_DQS_MODE_R

Bits 0:2 - the output spi_dqs is delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the spi_clk. Can be configured in CONF state.

pub fn d_cd_mode(&self) -> D_CD_MODE_R

Bits 3:5 - the output spi_cd is delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the spi_clk. Can be configured in CONF state.

pub fn d_de_mode(&self) -> D_DE_MODE_R

Bits 6:8 - the output spi_de is delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the spi_clk. Can be configured in CONF state.

pub fn d_hsync_mode(&self) -> D_HSYNC_MODE_R

Bits 9:11 - the output spi_hsync is delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the spi_clk. Can be configured in CONF state.

pub fn d_vsync_mode(&self) -> D_VSYNC_MODE_R

Bits 12:14 - the output spi_vsync is delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the spi_clk. Can be configured in CONF state.

pub fn de_idle_pol(&self) -> DE_IDLE_POL_R

Bit 15 - It is the idle value of spi_de.

pub fn hs_blank_en(&self) -> HS_BLANK_EN_R

Bit 16 - 1: The pulse of spi_hsync is out in vertical blanking lines in seg-trans or one trans. 0: spi_hsync pulse is valid only in active region lines in seg-trans.