Module esp32s2::spi0::slave

SPI slave control register

Structs

SLAVE_SPEC

SPI slave control register

Type Aliases

INT_DMA_SEG_TRANS_EN_R

Field INT_DMA_SEG_TRANS_EN reader - SPI_DMA_SEG_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_DMA_SEG_TRANS_EN_W

Field INT_DMA_SEG_TRANS_EN writer - SPI_DMA_SEG_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_RD_BUF_DONE_EN_R

Field INT_RD_BUF_DONE_EN reader - SPI_SLV_RD_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_RD_BUF_DONE_EN_W

Field INT_RD_BUF_DONE_EN writer - SPI_SLV_RD_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_RD_DMA_DONE_EN_R

Field INT_RD_DMA_DONE_EN reader - SPI_SLV_RD_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_RD_DMA_DONE_EN_W

Field INT_RD_DMA_DONE_EN writer - SPI_SLV_RD_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_TRANS_DONE_EN_R

Field INT_TRANS_DONE_EN reader - SPI_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_TRANS_DONE_EN_W

Field INT_TRANS_DONE_EN writer - SPI_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_WR_BUF_DONE_EN_R

Field INT_WR_BUF_DONE_EN reader - SPI_SLV_WR_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_WR_BUF_DONE_EN_W

Field INT_WR_BUF_DONE_EN writer - SPI_SLV_WR_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_WR_DMA_DONE_EN_R

Field INT_WR_DMA_DONE_EN reader - SPI_SLV_WR_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

INT_WR_DMA_DONE_EN_W

Field INT_WR_DMA_DONE_EN writer - SPI_SLV_WR_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

MODE_R

Field MODE reader - Set SPI work mode. 1: slave mode 0: master mode.

MODE_W

Field MODE writer - Set SPI work mode. 1: slave mode 0: master mode.

R

Register SLAVE reader

SEG_MAGIC_ERR_INT_EN_R

Field SEG_MAGIC_ERR_INT_EN reader - 1: Enable seg magic value error interrupt. 0: Others. Can be configured in CONF state.

SEG_MAGIC_ERR_INT_EN_W

Field SEG_MAGIC_ERR_INT_EN writer - 1: Enable seg magic value error interrupt. 0: Others. Can be configured in CONF state.

SOFT_RESET_R

Field SOFT_RESET reader - Software reset enable, reset the spi clock line cs line and data lines. Can be configured in CONF state.

SOFT_RESET_W

Field SOFT_RESET writer - Software reset enable, reset the spi clock line cs line and data lines. Can be configured in CONF state.

TRANS_CNT_R

Field TRANS_CNT reader - The operations counter in both the master mode and the slave mode.

TRANS_DONE_AUTO_CLR_EN_R

Field TRANS_DONE_AUTO_CLR_EN reader - SPI_TRANS_DONE auto clear enable, clear it 3 apb cycles after the pos edge of SPI_TRANS_DONE. 0:disable. 1: enable. Can be configured in CONF state.

TRANS_DONE_AUTO_CLR_EN_W

Field TRANS_DONE_AUTO_CLR_EN writer - SPI_TRANS_DONE auto clear enable, clear it 3 apb cycles after the pos edge of SPI_TRANS_DONE. 0:disable. 1: enable. Can be configured in CONF state.

TRANS_DONE_R

Field TRANS_DONE reader - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode. Can not be changed by CONF_buf.

TRANS_DONE_W

Field TRANS_DONE writer - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode. Can not be changed by CONF_buf.

W

Register SLAVE writer