Type Alias esp32s2::spi0::slave::W

pub type W = W<SLAVE_SPEC>;

Register SLAVE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn trans_done(&mut self) -> TRANS_DONE_W<'_, SLAVE_SPEC>

Bit 4 - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode. Can not be changed by CONF_buf.

pub fn int_rd_buf_done_en(&mut self) -> INT_RD_BUF_DONE_EN_W<'_, SLAVE_SPEC>

Bit 5 - SPI_SLV_RD_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn int_wr_buf_done_en(&mut self) -> INT_WR_BUF_DONE_EN_W<'_, SLAVE_SPEC>

Bit 6 - SPI_SLV_WR_BUF_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn int_rd_dma_done_en(&mut self) -> INT_RD_DMA_DONE_EN_W<'_, SLAVE_SPEC>

Bit 7 - SPI_SLV_RD_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn int_wr_dma_done_en(&mut self) -> INT_WR_DMA_DONE_EN_W<'_, SLAVE_SPEC>

Bit 8 - SPI_SLV_WR_DMA_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn int_trans_done_en(&mut self) -> INT_TRANS_DONE_EN_W<'_, SLAVE_SPEC>

Bit 9 - SPI_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn int_dma_seg_trans_en(&mut self) -> INT_DMA_SEG_TRANS_EN_W<'_, SLAVE_SPEC>

Bit 10 - SPI_DMA_SEG_TRANS_DONE Interrupt enable. 1: enable 0: disable. Can be configured in CONF state.

pub fn seg_magic_err_int_en(&mut self) -> SEG_MAGIC_ERR_INT_EN_W<'_, SLAVE_SPEC>

Bit 11 - 1: Enable seg magic value error interrupt. 0: Others. Can be configured in CONF state.

pub fn trans_done_auto_clr_en( &mut self ) -> TRANS_DONE_AUTO_CLR_EN_W<'_, SLAVE_SPEC>

Bit 29 - SPI_TRANS_DONE auto clear enable, clear it 3 apb cycles after the pos edge of SPI_TRANS_DONE. 0:disable. 1: enable. Can be configured in CONF state.

pub fn mode(&mut self) -> MODE_W<'_, SLAVE_SPEC>

Bit 30 - Set SPI work mode. 1: slave mode 0: master mode.

pub fn soft_reset(&mut self) -> SOFT_RESET_W<'_, SLAVE_SPEC>

Bit 31 - Software reset enable, reset the spi clock line cs line and data lines. Can be configured in CONF state.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual