Module esp32s2::spi0::slave1

SPI slave control register 1

Structs

SLAVE1_SPEC

SPI slave control register 1

Type Aliases

R

Register SLAVE1 reader

SLV_ADDR_ERR_CLR_R

Field SLV_ADDR_ERR_CLR reader - 1: Clear SPI_SLV_ADDR_ERR. 0: not valid. Can be changed by CONF_buf.

SLV_ADDR_ERR_CLR_W

Field SLV_ADDR_ERR_CLR writer - 1: Clear SPI_SLV_ADDR_ERR. 0: not valid. Can be changed by CONF_buf.

SLV_ADDR_ERR_R

Field SLV_ADDR_ERR reader - 1: The address value of the last SPI transfer is not supported by SPI slave. 0: The address value is supported or no address value is received.

SLV_CMD_ERR_CLR_R

Field SLV_CMD_ERR_CLR reader - 1: Clear SPI_SLV_CMD_ERR. 0: not valid. Can be changed by CONF_buf.

SLV_CMD_ERR_CLR_W

Field SLV_CMD_ERR_CLR writer - 1: Clear SPI_SLV_CMD_ERR. 0: not valid. Can be changed by CONF_buf.

SLV_CMD_ERR_R

Field SLV_CMD_ERR reader - 1: The command value of the last SPI transfer is not supported by SPI slave. 0: The command value is supported or no command value is received.

SLV_LAST_ADDR_R

Field SLV_LAST_ADDR reader - In the slave mode it is the value of address.

SLV_LAST_ADDR_W

Field SLV_LAST_ADDR writer - In the slave mode it is the value of address.

SLV_LAST_COMMAND_R

Field SLV_LAST_COMMAND reader - In the slave mode it is the value of command.

SLV_LAST_COMMAND_W

Field SLV_LAST_COMMAND writer - In the slave mode it is the value of command.

SLV_NO_QPI_EN_R

Field SLV_NO_QPI_EN reader - 1: spi slave QPI mode is not supported. 0: spi slave QPI mode is supported.

SLV_NO_QPI_EN_W

Field SLV_NO_QPI_EN writer - 1: spi slave QPI mode is not supported. 0: spi slave QPI mode is supported.

SLV_WR_DMA_DONE_R

Field SLV_WR_DMA_DONE reader - The interrupt raw bit for the completion of dma write operation in the slave mode. Can not be changed by CONF_buf.

SLV_WR_DMA_DONE_W

Field SLV_WR_DMA_DONE writer - The interrupt raw bit for the completion of dma write operation in the slave mode. Can not be changed by CONF_buf.

W

Register SLAVE1 writer