Type Alias esp32s2::spi0::slave1::R

pub type R = R<SLAVE1_SPEC>;

Register SLAVE1 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn slv_addr_err_clr(&self) -> SLV_ADDR_ERR_CLR_R

Bit 10 - 1: Clear SPI_SLV_ADDR_ERR. 0: not valid. Can be changed by CONF_buf.

pub fn slv_cmd_err_clr(&self) -> SLV_CMD_ERR_CLR_R

Bit 11 - 1: Clear SPI_SLV_CMD_ERR. 0: not valid. Can be changed by CONF_buf.

pub fn slv_no_qpi_en(&self) -> SLV_NO_QPI_EN_R

Bit 12 - 1: spi slave QPI mode is not supported. 0: spi slave QPI mode is supported.

pub fn slv_addr_err(&self) -> SLV_ADDR_ERR_R

Bit 13 - 1: The address value of the last SPI transfer is not supported by SPI slave. 0: The address value is supported or no address value is received.

pub fn slv_cmd_err(&self) -> SLV_CMD_ERR_R

Bit 14 - 1: The command value of the last SPI transfer is not supported by SPI slave. 0: The command value is supported or no command value is received.

pub fn slv_wr_dma_done(&self) -> SLV_WR_DMA_DONE_R

Bit 15 - The interrupt raw bit for the completion of dma write operation in the slave mode. Can not be changed by CONF_buf.

pub fn slv_last_command(&self) -> SLV_LAST_COMMAND_R

Bits 16:23 - In the slave mode it is the value of command.

pub fn slv_last_addr(&self) -> SLV_LAST_ADDR_R

Bits 24:31 - In the slave mode it is the value of address.