Type Alias esp32s2::spi0::slave1::W

pub type W = W<SLAVE1_SPEC>;

Register SLAVE1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn slv_addr_err_clr(&mut self) -> SLV_ADDR_ERR_CLR_W<'_, SLAVE1_SPEC>

Bit 10 - 1: Clear SPI_SLV_ADDR_ERR. 0: not valid. Can be changed by CONF_buf.

pub fn slv_cmd_err_clr(&mut self) -> SLV_CMD_ERR_CLR_W<'_, SLAVE1_SPEC>

Bit 11 - 1: Clear SPI_SLV_CMD_ERR. 0: not valid. Can be changed by CONF_buf.

pub fn slv_no_qpi_en(&mut self) -> SLV_NO_QPI_EN_W<'_, SLAVE1_SPEC>

Bit 12 - 1: spi slave QPI mode is not supported. 0: spi slave QPI mode is supported.

pub fn slv_wr_dma_done(&mut self) -> SLV_WR_DMA_DONE_W<'_, SLAVE1_SPEC>

Bit 15 - The interrupt raw bit for the completion of dma write operation in the slave mode. Can not be changed by CONF_buf.

pub fn slv_last_command(&mut self) -> SLV_LAST_COMMAND_W<'_, SLAVE1_SPEC>

Bits 16:23 - In the slave mode it is the value of command.

pub fn slv_last_addr(&mut self) -> SLV_LAST_ADDR_W<'_, SLAVE1_SPEC>

Bits 24:31 - In the slave mode it is the value of address.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual