Type Alias esp32s2::spi0::slv_rd_byte::W

pub type W = W<SLV_RD_BYTE_SPEC>;

Register SLV_RD_BYTE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn slv_data_bytelen(&mut self) -> SLV_DATA_BYTELEN_W<'_, SLV_RD_BYTE_SPEC>

Bits 0:19 - The full-duplex or half-duplex data byte length of the last SPI transfer in slave mode. In half-duplex mode, this value is controlled by bits [23:20].

pub fn slv_rddma_bytelen_en( &mut self ) -> SLV_RDDMA_BYTELEN_EN_W<'_, SLV_RD_BYTE_SPEC>

Bit 20 - 1: SPI_SLV_DATA_BYTELEN stores data byte length of master-read-slave data length in DMA controlled mode(Rd_DMA). 0: others

pub fn slv_wrdma_bytelen_en( &mut self ) -> SLV_WRDMA_BYTELEN_EN_W<'_, SLV_RD_BYTE_SPEC>

Bit 21 - 1: SPI_SLV_DATA_BYTELEN stores data byte length of master-write-to-slave data length in DMA controlled mode(Wr_DMA). 0: others

pub fn slv_rdbuf_bytelen_en( &mut self ) -> SLV_RDBUF_BYTELEN_EN_W<'_, SLV_RD_BYTE_SPEC>

Bit 22 - 1: SPI_SLV_DATA_BYTELEN stores data byte length of master-read-slave data length in CPU controlled mode(Rd_BUF). 0: others

pub fn slv_wrbuf_bytelen_en( &mut self ) -> SLV_WRBUF_BYTELEN_EN_W<'_, SLV_RD_BYTE_SPEC>

Bit 23 - 1: SPI_SLV_DATA_BYTELEN stores data byte length of master-write-to-slave data length in CPU controlled mode(Wr_BUF). 0: others

pub fn dma_seg_magic_value( &mut self ) -> DMA_SEG_MAGIC_VALUE_W<'_, SLV_RD_BYTE_SPEC>

Bits 24:27 - The magic value of BM table in master DMA seg-trans.

pub fn slv_rd_dma_done(&mut self) -> SLV_RD_DMA_DONE_W<'_, SLV_RD_BYTE_SPEC>

Bit 30 - The interrupt raw bit for the completion of Rd-DMA operation in the slave mode. Can not be changed by CONF_buf.

pub fn usr_conf(&mut self) -> USR_CONF_W<'_, SLV_RD_BYTE_SPEC>

Bit 31 - 1: Enable the DMA CONF phase of current seg-trans operation, which means seg-trans will start. 0: This is not seg-trans mode.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual