Type Alias esp32s2::system::perip_clk_en0::W

pub type W = W<PERIP_CLK_EN0_SPEC>;

Register PERIP_CLK_EN0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn timers_clk_en(&mut self) -> TIMERS_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 0 - Set this bit to enable clock of timers.

pub fn spi01_clk_en(&mut self) -> SPI01_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 1 - Set this bit to enable clock of SPI0 and SPI1.

pub fn uart_clk_en(&mut self) -> UART_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 2 - Set this bit to enable clock of UART0.

pub fn wdg_clk_en(&mut self) -> WDG_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 3 - Set this bit to enable clock of WDG.

pub fn i2s0_clk_en(&mut self) -> I2S0_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 4 - Set this bit to enable clock of I2S0.

pub fn uart1_clk_en(&mut self) -> UART1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 5 - Set this bit to enable clock of UART1.

pub fn spi2_clk_en(&mut self) -> SPI2_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 6 - Set this bit to enable clock of SPI2.

pub fn i2c_ext0_clk_en(&mut self) -> I2C_EXT0_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 7 - Set this bit to enable clock of I2C EXT0.

pub fn uhci0_clk_en(&mut self) -> UHCI0_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 8 - Set this bit to enable clock of UHCI0.

pub fn rmt_clk_en(&mut self) -> RMT_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 9 - Set this bit to enable clock of remote controller.

pub fn pcnt_clk_en(&mut self) -> PCNT_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 10 - Set this bit to enable clock of pulse count.

pub fn ledc_clk_en(&mut self) -> LEDC_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 11 - Set this bit to enable clock of LED PWM.

pub fn uhci1_clk_en(&mut self) -> UHCI1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 12 - Set this bit to enable clock of UHCI1.

pub fn timergroup_clk_en( &mut self ) -> TIMERGROUP_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 13 - Set this bit to enable clock of timer group0.

pub fn efuse_clk_en(&mut self) -> EFUSE_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 14 - Set this bit to enable clock of eFuse.

pub fn timergroup1_clk_en( &mut self ) -> TIMERGROUP1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 15 - Set this bit to enable clock of timer group1.

pub fn spi3_clk_en(&mut self) -> SPI3_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 16 - Set this bit to enable clock of SPI3.

pub fn pwm0_clk_en(&mut self) -> PWM0_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 17 - Set this bit to enable clock of PWM0.

pub fn i2c_ext1_clk_en(&mut self) -> I2C_EXT1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 18 - Set this bit to enable clock of I2C EXT1.

pub fn twai_clk_en(&mut self) -> TWAI_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 19 - Set this bit to enable clock of CAN.

pub fn pwm1_clk_en(&mut self) -> PWM1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 20 - Set this bit to enable clock of PWM1.

pub fn i2s1_clk_en(&mut self) -> I2S1_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 21 - Set this bit to enable clock of I2S1.

pub fn spi2_dma_clk_en(&mut self) -> SPI2_DMA_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 22 - Set this bit to enable clock of SPI2 DMA.

pub fn usb_clk_en(&mut self) -> USB_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 23 - Set this bit to enable clock of USB.

pub fn uart_mem_clk_en(&mut self) -> UART_MEM_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 24 - Set this bit to enable clock of UART memory.

pub fn pwm2_clk_en(&mut self) -> PWM2_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 25 - Set this bit to enable clock of PWM2.

pub fn pwm3_clk_en(&mut self) -> PWM3_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 26 - Set this bit to enable clock of PWM3.

pub fn spi3_dma_clk_en(&mut self) -> SPI3_DMA_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 27 - Set this bit to enable clock of SPI3 DMA.

pub fn apb_saradc_clk_en( &mut self ) -> APB_SARADC_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 28 - Set this bit to enable clock of SAR ADC.

pub fn systimer_clk_en(&mut self) -> SYSTIMER_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 29 - Set this bit to enable clock of system timer.

pub fn adc2_arb_clk_en(&mut self) -> ADC2_ARB_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 30 - Set this bit to enable clock of aribiter of ADC2.

pub fn spi4_clk_en(&mut self) -> SPI4_CLK_EN_W<'_, PERIP_CLK_EN0_SPEC>

Bit 31 - Set this bit to enable clock of SPI4.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual