Type Alias esp32s2::timg0::tconfig::W

pub type W = W<TCONFIG_SPEC>;

Register T%sCONFIG writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn use_xtal(&mut self) -> USE_XTAL_W<'_, TCONFIG_SPEC>

Bit 9 - 1: Use XTAL_CLK as the source clock of timer group. 0: Use APB_CLK as the source clock of timer group.

pub fn alarm_en(&mut self) -> ALARM_EN_W<'_, TCONFIG_SPEC>

Bit 10 - When set, the alarm is enabled. This bit is automatically cleared once an alarm occurs.

pub fn level_int_en(&mut self) -> LEVEL_INT_EN_W<'_, TCONFIG_SPEC>

Bit 11 - When set, an alarm will generate a level type interrupt.

pub fn edge_int_en(&mut self) -> EDGE_INT_EN_W<'_, TCONFIG_SPEC>

Bit 12 - When set, an alarm will generate an edge type interrupt.

pub fn divider(&mut self) -> DIVIDER_W<'_, TCONFIG_SPEC>

Bits 13:28 - Timer %s clock (T%s_clk) prescaler value.

pub fn autoreload(&mut self) -> AUTORELOAD_W<'_, TCONFIG_SPEC>

Bit 29 - When set, timer %s auto-reload at alarm is enabled.

pub fn increase(&mut self) -> INCREASE_W<'_, TCONFIG_SPEC>

Bit 30 - When set, the timer %s time-base counter will increment every clock tick. When cleared, the timer %s time-base counter will decrement.

pub fn en(&mut self) -> EN_W<'_, TCONFIG_SPEC>

Bit 31 - When set, the timer %s time-base counter is enabled.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual