Module esp32s2::twai0::status

Status register

Structs

STATUS_SPEC

Status register

Type Aliases

BUS_OFF_ST_R

Field BUS_OFF_ST reader - 1: In bus-off status, the TWAI Controller is no longer involved in bus activities.

ERR_ST_R

Field ERR_ST reader - 1: At least one of the RX/TX error counter has reached or exceeded the value set in register TWAI_ERR_WARNING_LIMIT_REG.

MISS_ST_R

Field MISS_ST reader - This bit reflects whether the data packet in the RX FIFO is complete. 1: The current packet is missing; 0: The current packet is complete

OVERRUN_ST_R

Field OVERRUN_ST reader - 1: The RX FIFO is full and data overrun has occurred.

R

Register STATUS reader

RX_BUF_ST_R

Field RX_BUF_ST reader - 1: The data in the RX buffer is not empty, with at least one received data packet.

RX_ST_R

Field RX_ST reader - 1: The TWAI Controller is receiving a message from the bus.

TX_BUF_ST_R

Field TX_BUF_ST reader - 1: The TX buffer is empty, the CPU may write a message into it.

TX_COMPLETE_R

Field TX_COMPLETE reader - 1: The TWAI controller has successfully received a packet from the bus.

TX_ST_R

Field TX_ST reader - 1: The TWAI Controller is transmitting a message to the bus.