Module esp32s2::uart0::status

UART status register

Structs

STATUS_SPEC

UART status register

Type Aliases

CTSN_R

Field CTSN reader - This register represents the level of the internal UART CTS signal.

DSRN_R

Field DSRN reader - This register represents the level of the internal UART DSR signal.

DTRN_R

Field DTRN reader - This bit represents the level of the internal UART DTR signal.

R

Register STATUS reader

RTSN_R

Field RTSN reader - This bit represents the level of the internal UART RTS signal.

RXD_R

Field RXD reader - This register represents the level of the internal UART RXD signal.

RXFIFO_CNT_R

Field RXFIFO_CNT reader - Stores the number of valid data bytes in RX FIFO.

TXD_R

Field TXD reader - This bit represents the level of the internal UART TXD signal.

TXFIFO_CNT_R

Field TXFIFO_CNT reader - Stores the number of data bytes in TX FIFO.