Type Alias esp32s2::uhci0::conf0::W

pub type W = W<CONF0_SPEC>;

Register CONF0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn in_rst(&mut self) -> IN_RST_W<'_, CONF0_SPEC>

Bit 0 - Set this bit to reset in DMA FSM.

pub fn out_rst(&mut self) -> OUT_RST_W<'_, CONF0_SPEC>

Bit 1 - Set this bit to reset out DMA FSM.

pub fn ahbm_fifo_rst(&mut self) -> AHBM_FIFO_RST_W<'_, CONF0_SPEC>

Bit 2 - Set this bit to reset AHB interface cmdFIFO of DMA.

pub fn ahbm_rst(&mut self) -> AHBM_RST_W<'_, CONF0_SPEC>

Bit 3 - Set this bit to reset AHB interface of DMA.

pub fn in_loop_test(&mut self) -> IN_LOOP_TEST_W<'_, CONF0_SPEC>

Bit 4 - Reserved.

pub fn out_loop_test(&mut self) -> OUT_LOOP_TEST_W<'_, CONF0_SPEC>

Bit 5 - Reserved.

pub fn out_auto_wrback(&mut self) -> OUT_AUTO_WRBACK_W<'_, CONF0_SPEC>

Bit 6 - Set this bit to enable automatic outlink writeback when all the data in TX FIFO has been transmitted.

pub fn out_no_restart_clr(&mut self) -> OUT_NO_RESTART_CLR_W<'_, CONF0_SPEC>

Bit 7 - Reserved.

pub fn out_eof_mode(&mut self) -> OUT_EOF_MODE_W<'_, CONF0_SPEC>

Bit 8 - This register is used to specify the generation mode of UHCI_OUT_EOF_INT interrupt. 1: When DMA has popped all data from FIFO. 0: When AHB has pushed all data to FIFO.

pub fn uart0_ce(&mut self) -> UART0_CE_W<'_, CONF0_SPEC>

Bit 9 - Set this bit to link up UHCI and UART0.

pub fn uart1_ce(&mut self) -> UART1_CE_W<'_, CONF0_SPEC>

Bit 10 - Set this bit to link up UHCI and UART1.

pub fn outdscr_burst_en(&mut self) -> OUTDSCR_BURST_EN_W<'_, CONF0_SPEC>

Bit 12 - This register is used to specify DMA transmit descriptor transfer mode. 1: burst mode. 0: byte mode.

pub fn indscr_burst_en(&mut self) -> INDSCR_BURST_EN_W<'_, CONF0_SPEC>

Bit 13 - This register is used to specify DMA receive descriptor transfer mode. 1: burst mode. 0: byte mode.

pub fn mem_trans_en(&mut self) -> MEM_TRANS_EN_W<'_, CONF0_SPEC>

Bit 15 - 1: UHCI transmitted data would be write back into DMA INFIFO.

pub fn seper_en(&mut self) -> SEPER_EN_W<'_, CONF0_SPEC>

Bit 16 - Set this bit to separate the data frame using a special character.

pub fn head_en(&mut self) -> HEAD_EN_W<'_, CONF0_SPEC>

Bit 17 - Set this bit to encode the data packet with a formatting header.

pub fn crc_rec_en(&mut self) -> CRC_REC_EN_W<'_, CONF0_SPEC>

Bit 18 - Set this bit to enable UHCI to receive the 16 bit CRC.

pub fn uart_idle_eof_en(&mut self) -> UART_IDLE_EOF_EN_W<'_, CONF0_SPEC>

Bit 19 - If this bit is set to 1, UHCI will end the payload receiving process when UART has been in idle state.

pub fn len_eof_en(&mut self) -> LEN_EOF_EN_W<'_, CONF0_SPEC>

Bit 20 - If this bit is set to 1, UHCI decoder stops receiving payload data when the number of received data bytes has reached the specified value. The value is payload length indicated by UCHI packet header when UHCI_HEAD_EN is 1 or the value is a configuration value when UHCI_HEAD_EN is 0. If this bit is set to 0, UHCI decoder stops receiving payload data upon receiving 0xC0.

pub fn encode_crc_en(&mut self) -> ENCODE_CRC_EN_W<'_, CONF0_SPEC>

Bit 21 - Set this bit to enable data integrity checking by appending a 16 bit CCITT-CRC to the end of the payload.

pub fn clk_en(&mut self) -> CLK_EN_W<'_, CONF0_SPEC>

Bit 22 - 1: Force clock on for registers. 0: Support clock only when application writes registers.

pub fn uart_rx_brk_eof_en(&mut self) -> UART_RX_BRK_EOF_EN_W<'_, CONF0_SPEC>

Bit 23 - If this bit is set to 1, UHCI stops receiving payload data when a NULL frame is received by UART.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual